% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_build_ignore}
\alias{add_build_ignore}
\title{Add build ignore}
\usage{
add_build_ignore(build_ignore_ls)
}
\arguments{
\item{build_ignore_ls}{Build ignore (a list)}
}
\description{
add_build_ignore() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add build ignore. Function argument build_ignore_ls specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
