% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_fns_dmt_tb}
\alias{add_fns_dmt_tb}
\title{Add functions documentation tibble}
\usage{
add_fns_dmt_tb(
  pkg_setup_ls,
  append_1L_lgl = T,
  dv_url_pfx_1L_chr = character(0),
  fns_env_ls = NULL,
  inc_methods_1L_lgl = F,
  key_1L_chr = NULL
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{append_1L_lgl}{Append (a logical vector of length one), Default: T}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: character(0)}

\item{fns_env_ls}{Functions (a list of environments), Default: NULL}

\item{inc_methods_1L_lgl}{Include methods (a logical vector of length one), Default: F}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Package setup (a list)
}
\description{
add_fns_dmt_tb() is an Add function that updates an object by adding new values to new or empty fields Specifically, this function implements an algorithm to add functions documentation tibble. The function returns Package setup (a list).
}
\keyword{internal}
