% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_indefartls_to_phrases}
\alias{add_indefartls_to_phrases}
\title{Add indefinite articles to phrases}
\usage{
add_indefartls_to_phrases(
  abbreviated_phrase_1L_chr,
  abbreviations_lup = NULL,
  ignore_phrs_not_in_lup_1L_lgl = T
)
}
\arguments{
\item{abbreviated_phrase_1L_chr}{Abbreviated phrase (a character vector of length one)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{ignore_phrs_not_in_lup_1L_lgl}{Ignore phrases not in lookup table (a logical vector of length one), Default: T}
}
\value{
Phrases (a character vector)
}
\description{
add_indefartls_to_phrases() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add indefinite articles to phrases. Function argument abbreviated_phrase_1L_chr specifies the object to be updated. The function returns phrases (a character vector).
}
\keyword{internal}
