% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_new_cls_pts}
\alias{add_new_cls_pts}
\title{Add new class prototypes}
\usage{
add_new_cls_pts(pkg_setup_ls, addl_cls_pts_tb = NULL)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{addl_cls_pts_tb}{Additional class prototypes (a tibble), Default: NULL}
}
\value{
Package setup (a list)
}
\description{
add_new_cls_pts() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add new class prototypes. Function argument pkg_setup_ls specifies the object to be updated. The function returns Package setup (a list).
}
