% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_plurals_to_abbr_lup_tb}
\alias{add_plurals_to_abbr_lup_tb}
\title{Add plurals to abbreviation lookup table}
\usage{
add_plurals_to_abbr_lup_tb(
  abbr_tb,
  no_plural_chr_vec = NA_character_,
  custom_plural_ls = NULL
)
}
\arguments{
\item{abbr_tb}{Abbreviation (a tibble)}

\item{no_plural_chr_vec}{No plural (a character vector), Default: 'NA'}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}
}
\value{
Abbreviation (a tibble)
}
\description{
add_plurals_to_abbr_lup_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add plurals to an abbreviation lookup table. Function argument abbr_tb specifies the object to be updated. The function returns an abbreviation (a tibble).
}
\keyword{internal}
