% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_rows_to_fn_type_lup}
\alias{add_rows_to_fn_type_lup}
\title{Add rows to function type}
\usage{
add_rows_to_fn_type_lup(
  fn_types_lup = make_fn_type_lup(),
  fn_type_nm_chr = NA_character_,
  fn_type_desc_chr = NA_character_,
  first_arg_desc_chr = NA_character_,
  second_arg_desc_chr = NA_character_,
  is_generic_lgl = F,
  is_method_lgl = F
)
}
\arguments{
\item{fn_types_lup}{Function type lookup table (a tibble), Default: make_fn_type_lup()}

\item{fn_type_nm_chr}{Function type name (a character vector), Default: 'NA'}

\item{fn_type_desc_chr}{Function type description (a character vector), Default: 'NA'}

\item{first_arg_desc_chr}{First argument description (a character vector), Default: 'NA'}

\item{second_arg_desc_chr}{Second argument description (a character vector), Default: 'NA'}

\item{is_generic_lgl}{Is generic (a logical vector), Default: F}

\item{is_method_lgl}{Is method (a logical vector), Default: F}
}
\value{
Updated function type lookup table (a tibble)
}
\description{
add_rows_to_fn_type_lup() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add rows to function type lookup table. Function argument fn_types_lup specifies the object to be updated. The function returns Updated function type lookup table (a tibble).
}
\keyword{internal}
