% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_addl_pkgs_ls.R
\name{addl_pkgs_ls<--ready4fun_pkg_setup_two}
\alias{addl_pkgs_ls<--ready4fun_pkg_setup_two}
\alias{addl_pkgs_ls<-,ready4fun_pkg_setup_two-method}
\title{addl_pkgs_ls<-}
\usage{
\S4method{addl_pkgs_ls}{ready4fun_pkg_setup_two}(x) <- value
}
\arguments{
\item{x}{An object of class ready4fun_pkg_setup_two}
}
\description{
Set the value of the slot addl_pkgs_ls for S4 objects of class ready4fun_pkg_setup_two
}
