% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{analyse}
\alias{analyse}
\title{Analyse}
\usage{
analyse(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
analyse() is an Analyse generic that runs an analysis or analyses using specified data and parameters.
}
