% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_assert.R
\name{assert_does_not_match_terms}
\alias{assert_does_not_match_terms}
\title{Assert does not match terms}
\usage{
assert_does_not_match_terms(input_chr_vec, exclude_if_match_chr_vec)
}
\arguments{
\item{input_chr_vec}{Input (a character vector)}

\item{exclude_if_match_chr_vec}{Exclude if match (a character vector)}
}
\description{
assert_does_not_match_terms() is an Assert function that validates that an object conforms to required condition(s). If the object does not meet all required conditions, program execution will be stopped and an error message provided. Specifically, this function implements an algorithm to assert does not match terms. Function argument input_chr_vec specifies the object on which assert validation checks are to be performed. Argument exclude_if_match_chr_vec provides the object containing values used for validation tests.NA
}
\keyword{internal}
