% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_assert.R
\name{assert_inp_does_not_match_terms}
\alias{assert_inp_does_not_match_terms}
\title{assert input does not match terms}
\usage{
assert_inp_does_not_match_terms(input_chr, exclude_if_match_chr)
}
\arguments{
\item{input_chr}{Input (a character vector)}

\item{exclude_if_match_chr}{Exclude if match (a character vector)}
}
\description{
assert_inp_does_not_match_terms() is an Assert function that validates that an object conforms to required condition(s). If the object does not meet all required conditions, program execution will be stopped and an error message provided. Specifically, this function implements an algorithm to assert input does not match terms. Function argument input_chr specifies the object on which assert validation checks are to be performed. Argument exclude_if_match_chr provides the object containing values used for validation tests. The function is called for its side effects and does not return a value.
}
\keyword{internal}
