% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_authorClasses.R
\name{authorClasses}
\alias{authorClasses}
\alias{authorClasses.ready4fun_manifest}
\alias{authorClasses,ready4fun_manifest-method}
\title{AuthorClasses}
\usage{
authorClasses(x, ...)

authorClasses.ready4fun_manifest(
  x,
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)

\S4method{authorClasses}{ready4fun_manifest}(
  x,
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{x}{An instance of ready4 S3 class for encapsulating the metadata required for package set-up.}

\item{...}{Additional arguments (an additional arguments)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{self_serve_fn_ls}{Self serve (a list of functions), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\description{
authorClasses() is an AuthorClasses function that authors and saves files necessary for creating and documenting classes. Specifically, this function implements an algorithm to authorclasses. The function is called for its side effects and does not return a value.

authorClasses.ready4fun_manifest() is an AuthorClasses method that authors and saves files necessary for creating and documenting classes. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
