% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_authorData.R
\name{authorData}
\alias{authorData}
\alias{authorData.ready4fun_manifest}
\alias{authorData,ready4fun_manifest-method}
\title{AuthorData}
\usage{
authorData(x, ...)

authorData.ready4fun_manifest(x)

\S4method{authorData}{ready4fun_manifest}(x)
}
\arguments{
\item{x}{An instance of ready4 S3 class for encapsulating the metadata required for package set-up.}

\item{...}{Additional arguments (an additional arguments)}
}
\value{
NA ()
}
\description{
authorData() is an AuthorData function that authors and saves files necessary for creating and documenting datasets. Specifically, this function implements an algorithm to authordata. The function is called for its side effects and does not return a value.

authorData.ready4fun_manifest() is an AuthorData method that authors and saves files necessary for creating and documenting datasets. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
