% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_authorFunctions.R
\name{authorFunctions}
\alias{authorFunctions}
\alias{authorFunctions.ready4fun_Manifest}
\alias{authorFunctions,ready4fun_Manifest-method}
\title{AuthorFunctions}
\usage{
authorFunctions(x, ...)

authorFunctions.ready4fun_Manifest(x, list_generics_1L_lgl = F)

\S4method{authorFunctions}{ready4fun_Manifest}(x, list_generics_1L_lgl = F)
}
\arguments{
\item{x}{An instance of}

\item{...}{Additional arguments (an additional arguments)}

\item{list_generics_1L_lgl}{List generics (a logical vector of length one), Default: F}
}
\description{
authorFunctions() is an AuthorFunctions function that authors and saves files necessary for creating and documenting functions. Specifically, this function implements an algorithm to authorfunctions. The function is called for its side effects and does not return a value.

authorFunctions.ready4fun_Manifest() is an AuthorFunctions function that authors and saves files necessary for creating and documenting functions. Specifically, this function implements an algorithm to authorfunctions.ready4fun manifest. The function is called for its side effects and does not return a value.
}
