% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_badges_lup.R
\docType{data}
\name{badges_lup}
\alias{badges_lup}
\title{ready4 badges lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 4 columns.
}
\source{
\url{https://readyforwhatsnext.github.io/readyforwhatsnext/}
}
\usage{
badges_lup
}
\description{
A lookup table to identify the appropriate text to insert in README files to represent different types of ready4 badges.
}
\details{
A tibble

\describe{
\item{badge_names_chr}{Badge names (a character vector)}
\item{label_names_chr}{Label names (a character vector)}
\item{colours_chr}{Colours (a character vector)}
\item{badges_chr}{Badges (a character vector)}
}
}
\keyword{datasets}
