% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_close.R
\name{close_open_sinks}
\alias{close_open_sinks}
\title{Close open sinks}
\usage{
close_open_sinks()
}
\description{
close_open_sinks() is a Close function that closes specified connections. Specifically, this function implements an algorithm to close open sinks. The function is called for its side effects and does not return a value.
}
\keyword{internal}
