% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit.ready4fun_abbreviations}
\alias{exhibit.ready4fun_abbreviations}
\alias{exhibit,ready4fun_abbreviations-method}
\alias{exhibit.ready4fun_functions}
\alias{exhibit,ready4fun_functions-method}
\alias{exhibit.ready4fun_objects}
\alias{exhibit,ready4fun_objects-method}
\title{Exhibit (print to console) features of a ready4 framework module (or sub-module)}
\usage{
exhibit.ready4fun_abbreviations(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  select_int = NA_integer_,
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{ready4fun_abbreviations}(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  select_int = NA_integer_,
  use_lbls_as_col_nms_1L_lgl = T
)

exhibit.ready4fun_functions(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  select_int = NA_integer_,
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{ready4fun_functions}(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  select_int = NA_integer_,
  use_lbls_as_col_nms_1L_lgl = T
)

exhibit.ready4fun_objects(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  select_int = NA_integer_,
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{ready4fun_objects}(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  select_int = NA_integer_,
  use_lbls_as_col_nms_1L_lgl = T
)
}
\arguments{
\item{x}{An instance of Object types lookup table.}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: NULL}

\item{mkdn_tbl_ref_1L_chr}{Markdown table reference (a character vector of length one), Default: NULL}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{select_int}{Select (an integer vector), Default: NA}

\item{use_lbls_as_col_nms_1L_lgl}{Use labels as column names (a logical vector of length one), Default: T}
}
\description{
exhibit.ready4fun_abbreviations() is an exhibit method that exhibits features of a class instance by printing to console. This method is implemented for the ready4 S3 class for tibble object lookup table of abbreviations. The function is called for its side effects and does not return a value.

exhibit.ready4fun_functions() is an exhibit method that exhibits features of a class instance by printing to console. This method is implemented for the Function types lookup table. The function is called for its side effects and does not return a value.

exhibit.ready4fun_objects() is an exhibit method that exhibits features of a class instance by printing to console. This method is implemented for the Object types lookup table. The function is called for its side effects and does not return a value.
}
