% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_fn_types_lup.R
\docType{data}
\name{fn_types_lup}
\alias{fn_types_lup}
\title{Function type lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65 rows and 6 columns.
}
\source{
\url{ready4-dev/ready4}
}
\usage{
fn_types_lup
}
\description{
A lookup table to find descriptions for different types of functions used within the ready4fun package suite.
}
\details{
A tibble

\describe{
\item{fn_type_nm_chr}{Function type name (a character vector)}
\item{fn_type_desc_chr}{Function type description (a character vector)}
\item{first_arg_desc_chr}{First argument description (a character vector)}
\item{second_arg_desc_chr}{Second argument description (a character vector)}
\item{is_generic_lgl}{Is generic (a logical vector)}
\item{is_method_lgl}{Is method (a logical vector)}
}
}
\keyword{datasets}
