% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_force.R
\name{force_instl_of_reqd_pkg}
\alias{force_instl_of_reqd_pkg}
\title{Force install of required package}
\usage{
force_instl_of_reqd_pkg(package_nm_1L_chr)
}
\arguments{
\item{package_nm_1L_chr}{Package name (a character vector of length one)}
}
\description{
force_instl_of_reqd_pkg() is a Force function that checks if a specified local or global environmental condition is met and if not, updates the specified environment to comply with the condition. Specifically, this function implements an algorithm to force an install of a required package.NA
}
\keyword{internal}
