% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_force.R
\name{force_req_pkg_install}
\alias{force_req_pkg_install}
\title{Force req package install}
\usage{
force_req_pkg_install(package_nm_chr)
}
\arguments{
\item{package_nm_chr}{Package name (a character vector of length 1)}
}
\description{
force_req_pkg_install() is a Force function that checks if a specified local or global environmental condition is met and if not, updates the specified environment to comply with the condition. Specifically, this function implements an algorithm to force req a package install. The function is called for its side effects and does not return a value.
}
