% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dev_pkg_nm}
\alias{get_dev_pkg_nm}
\title{Get development package name}
\usage{
get_dev_pkg_nm(path_to_pkg_rt_1L_chr = ".")
}
\arguments{
\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: '.'}
}
\value{
Development package name (a character vector of length one)
}
\description{
get_dev_pkg_nm() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get development package name. The function returns Development package name (a character vector of length one).
}
\keyword{internal}
