% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dev_pkg_nm_1L_chr}
\alias{get_dev_pkg_nm_1L_chr}
\title{Get dev package name 1L}
\usage{
get_dev_pkg_nm_1L_chr(path_to_pkg_rt_chr = ".")
}
\arguments{
\item{path_to_pkg_rt_chr}{Path to package rt (a character vector of length 1), Default: '.'}
}
\value{
Dev package name (a character vector of length 1)
}
\description{
get_dev_pkg_nm_1L_chr() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get dev a package name 1L. Function argument path_to_pkg_rt_chr specifies the where to look for the required object.The function returns dev package name (a character vector of length 1).
}
\keyword{internal}
