% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dv_fls_urls}
\alias{get_dv_fls_urls}
\title{Get dataverse files urls}
\usage{
get_dv_fls_urls(
  file_nms_chr,
  dv_ds_nm_1L_chr,
  dv_url_pfx_1L_chr = "https://dataverse.harvard.edu/api/access/datafile/"
)
}
\arguments{
\item{file_nms_chr}{File names (a character vector)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: 'https://dataverse.harvard.edu/api/access/datafile/'}
}
\value{
Urls (a character vector)
}
\description{
get_dv_fls_urls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get dataverse files urls. Function argument file_nms_chr specifies the where to look for the required object. The function returns Urls (a character vector).
}
