% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fn_args}
\alias{get_fn_args}
\title{Get function arguments}
\usage{
get_fn_args(fn)
}
\arguments{
\item{fn}{Function (a function)}
}
\value{
Function arguments (a character vector)
}
\description{
get_fn_args() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get function arguments. The function returns Function arguments (a character vector).
}
\keyword{internal}
