% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mthd_title}
\alias{get_mthd_title}
\title{Get method title}
\usage{
get_mthd_title(mthd_nm_1L_chr, pkg_nm_1L_chr = "ready4")
}
\arguments{
\item{mthd_nm_1L_chr}{Method name (a character vector of length one)}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'ready4'}
}
\value{
Method title (a character vector of length one)
}
\description{
get_mthd_title() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get method title. Function argument mthd_nm_1L_chr specifies the where to look for the required object. The function returns Method title (a character vector of length one).
}
\keyword{internal}
