% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_abbrs}
\alias{get_new_abbrs}
\title{Get new abbreviations}
\usage{
get_new_abbrs(
  pkg_setup_ls,
  classes_to_make_tb = NULL,
  inc_all_mthds_1L_lgl = T,
  fns_dmt_tb = NULL,
  paths_ls = make_fn_nms(),
  pkg_ds_ls_ls = NULL,
  transformations_chr = NULL,
  undocumented_fns_dir_chr = make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T),
  use_last_1L_int = NULL
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{classes_to_make_tb}{Classes to make (a tibble), Default: NULL}

\item{inc_all_mthds_1L_lgl}{Include all methods (a logical vector of length one), Default: T}

\item{fns_dmt_tb}{Functions documentation (a tibble), Default: NULL}

\item{paths_ls}{Paths (a list), Default: make_fn_nms()}

\item{pkg_ds_ls_ls}{Package dataset (a list of lists), Default: NULL}

\item{transformations_chr}{Transformations (a character vector), Default: NULL}

\item{undocumented_fns_dir_chr}{Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)}

\item{use_last_1L_int}{Use last (an integer vector of length one), Default: NULL}
}
\value{
New abbreviations (a character vector)
}
\description{
get_new_abbrs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get new abbreviations. Function argument pkg_setup_ls specifies the where to look for the required object. The function returns New abbreviations (a character vector).
}
\keyword{internal}
