% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_abbrs_cndts}
\alias{get_new_abbrs_cndts}
\title{get new abbreviations candidates}
\usage{
get_new_abbrs_cndts(
  text_chr,
  abbreviations_lup,
  drop_first_1L_lgl = F,
  use_last_1L_int = NULL,
  treat_as_words_chr = character(0)
)
}
\arguments{
\item{text_chr}{Text (a character vector)}

\item{abbreviations_lup}{Abbreviations (a lookup table)}

\item{drop_first_1L_lgl}{Drop first (a logical vector of length one), Default: F}

\item{use_last_1L_int}{Use last (an integer vector of length one), Default: NULL}

\item{treat_as_words_chr}{Treat as words (a character vector), Default: character(0)}
}
\value{
New abbreviations candidates (a character vector)
}
\description{
get_new_abbrs_cndts() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get new abbreviations candidates. Function argument text_chr specifies the where to look for the required object. The function returns New abbreviations candidates (a character vector).
}
\keyword{internal}
