% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_abbrvs}
\alias{get_new_abbrvs}
\title{Get new abbrvs}
\usage{
get_new_abbrvs(
  pkg_ds_ls_ls,
  pkg_setup_ls,
  abbreviations_lup = NULL,
  dv_ds_nm_1L_chr = "https://doi.org/10.7910/DVN/2Y9VF9",
  dv_url_pfx_1L_chr = NULL,
  fn_types_lup = NULL,
  inc_all_mthds_1L_lgl = T,
  key_1L_chr = NULL,
  object_type_lup = NULL,
  paths_ls = make_fn_nms(),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  undocumented_fns_dir_chr = make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)
)
}
\arguments{
\item{pkg_ds_ls_ls}{Package dataset (a list of lists)}

\item{pkg_setup_ls}{Package setup (a list)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'https://doi.org/10.7910/DVN/2Y9VF9'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{fn_types_lup}{Function type lookup table (a tibble), Default: NULL}

\item{inc_all_mthds_1L_lgl}{Include all methods (a logical vector of length one), Default: T}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{paths_ls}{Paths (a list), Default: make_fn_nms()}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{undocumented_fns_dir_chr}{Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)}
}
\value{
New abbrvs (a character vector)
}
\description{
get_new_abbrvs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get new abbrvs. Function argument pkg_ds_ls_ls specifies the where to look for the required object. The function returns New abbrvs (a character vector).
}
\keyword{internal}
