% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_abbrvs_cndts}
\alias{get_new_abbrvs_cndts}
\title{Get new abbrvs candidates}
\usage{
get_new_abbrvs_cndts(text_chr, drop_first_1L_lgl = F)
}
\arguments{
\item{text_chr}{Text (a character vector)}

\item{drop_first_1L_lgl}{Drop first (a logical vector of length one), Default: F}
}
\value{
New abbrvs candidates (a character vector)
}
\description{
get_new_abbrvs_cndts() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get new abbrvs candidates. Function argument text_chr specifies the where to look for the required object. The function returns New abbrvs candidates (a character vector).
}
\keyword{internal}
