% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_cls_pts}
\alias{get_new_cls_pts}
\title{Get new class prototypes}
\usage{
get_new_cls_pts(pkg_setup_ls)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}
}
\value{
New class prototypes (a character vector)
}
\description{
get_new_cls_pts() is a Get function that extracts data from an object Specifically, this function implements an algorithm to get new class prototypes. The function returns New class prototypes (a character vector).
}
\keyword{internal}
