% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_obj_type_new_cses}
\alias{get_obj_type_new_cses}
\title{Get object type new cases}
\usage{
get_obj_type_new_cses(
  updated_obj_type_lup,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  excluded_chr = NA_character_,
  key_1L_chr = deprecated(),
  old_obj_type_lup = NULL,
  server_1L_chr = deprecated()
)
}
\arguments{
\item{updated_obj_type_lup}{Updated object type (a lookup table)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{excluded_chr}{Excluded (a character vector), Default: 'NA'}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{old_obj_type_lup}{Old object type (a lookup table), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Object type lookup table new cases (a tibble)
}
\description{
get_obj_type_new_cses() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get object type new cases. The function returns Object type lookup table new cases (a tibble).
}
\keyword{internal}
