% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_obj_types}
\alias{get_obj_types}
\title{Get object types}
\usage{
get_obj_types(
  dv_nm_1L_chr = NA_character_,
  dv_ds_metadata_ls = list(list()),
  dv_ds_nm_1L_chr = NA_character_,
  dv_server_1L_chr = NA_character_,
  dv_url_pfx_1L_chr = NA_character_,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  type_1L_chr = c("submodule", "tibble"),
  what_1L_chr = c("seed", "all")
)
}
\arguments{
\item{dv_nm_1L_chr}{Dataverse name (a character vector of length one), Default: 'NA'}

\item{dv_ds_metadata_ls}{Dataverse dataset metadata (a list), Default: list(list())}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'NA'}

\item{dv_server_1L_chr}{Dataverse server (a character vector of length one), Default: 'NA'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: 'NA'}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("submodule", "tibble")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("seed", "all")}
}
\value{
Object type (a lookup table)
}
\description{
get_obj_types() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get object types. The function returns Object type (a lookup table).
}
\keyword{internal}
