% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_outp_obj_type_chr_vec}
\alias{get_outp_obj_type_chr_vec}
\title{Get output object type}
\usage{
get_outp_obj_type_chr_vec(fns_chr_vec)
}
\arguments{
\item{fns_chr_vec}{Functions (a character vector)}
}
\value{
Output object type (a character vector)
}
\description{
get_outp_obj_type_chr_vec() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get an output object type. Function argument fns_chr_vec specifies the where to look for the required object.The function returns an output object type (a character vector).
}
\keyword{internal}
