% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_r4_obj_slots_chr_vec}
\alias{get_r4_obj_slots_chr_vec}
\title{Get readyforwhatsnext S4 object slots}
\usage{
get_r4_obj_slots_chr_vec(fn_name_chr, package_chr = "")
}
\arguments{
\item{fn_name_chr}{Function name (a character vector of length 1)}

\item{package_chr}{Package (a character vector of length 1), Default: ''}
}
\value{
Slots (a character vector)
}
\description{
get_r4_obj_slots_chr_vec() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get a readyforwhatsnext S4 object slots. Function argument fn_name_chr specifies the where to look for the required object.The function returns slots (a character vector).
}
\keyword{internal}
