% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_rds_from_dv}
\alias{get_rds_from_dv}
\title{Get rds from dataverse}
\usage{
get_rds_from_dv(
  file_nm_1L_chr,
  dv_ds_nm_1L_chr = "https://doi.org/10.7910/DVN/2Y9VF9",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{file_nm_1L_chr}{File name (a character vector of length one)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'https://doi.org/10.7910/DVN/2Y9VF9'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Model (a model)
}
\description{
get_rds_from_dv() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get rds from dataverse. Function argument file_nm_1L_chr specifies the where to look for the required object. The function returns Model (a model).
}
\keyword{internal}
