% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_rds_from_pkg_dmt}
\alias{get_rds_from_pkg_dmt}
\title{Get rds from package documentation}
\usage{
get_rds_from_pkg_dmt(
  pkg_setup_ls = NULL,
  fl_nm_1L_chr,
  piggyback_to_1L_chr = character(0),
  piggyback_tag_1L_chr = "Documentation_0.0",
  piggyback_token_1L_chr = ""
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list), Default: NULL}

\item{fl_nm_1L_chr}{File name (a character vector of length one)}

\item{piggyback_to_1L_chr}{Piggyback to (a character vector of length one), Default: character(0)}

\item{piggyback_tag_1L_chr}{Piggyback tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{piggyback_token_1L_chr}{Piggyback token (a character vector of length one), Default: ''}
}
\value{
R object (an output object of multiple potential types)
}
\description{
get_rds_from_pkg_dmt() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get rds from package documentation. Function argument pkg_setup_ls specifies the where to look for the required object. The function returns R object (an output object of multiple potential types).
}
\keyword{internal}
