% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_return_obj_nm}
\alias{get_return_obj_nm}
\title{Get return object name}
\usage{
get_return_obj_nm(fn)
}
\arguments{
\item{fn}{Function (a function)}
}
\value{
Return (a character vector of length one)
}
\description{
get_return_obj_nm() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get return object name. The function returns Return (a character vector of length one).
}
\keyword{internal}
