% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_import.R
\name{import_xls_sheets}
\alias{import_xls_sheets}
\title{Import Excel workbook sheets}
\usage{
import_xls_sheets(range_1L_chr, sheet_names_chr, path_1L_chr)
}
\arguments{
\item{range_1L_chr}{Range (a character vector of length one)}

\item{sheet_names_chr}{Sheet names (a character vector)}

\item{path_1L_chr}{Path (a character vector of length one)}
}
\value{
Tibble list (a list of tibbles)
}
\description{
import_xls_sheets() is an Import function that reads a data object in its native format and converts it to an R object. Specifically, this function implements an algorithm to an import Excel workbook sheets.The function returns a tibble list (a list of tibbles).
}
\keyword{internal}
