% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_import.R
\name{import_xls_sheets_ls}
\alias{import_xls_sheets_ls}
\title{Import Excel workbook sheets}
\usage{
import_xls_sheets_ls(range_chr, sheet_names_chr_vec, path_chr)
}
\arguments{
\item{range_chr}{Range (a character vector of length 1)}

\item{sheet_names_chr_vec}{Sheet names (a character vector)}

\item{path_chr}{Path (a character vector of length 1)}
}
\value{
Tibble list (a list of tibbles)
}
\description{
import_xls_sheets_ls() is an Import function that reads a data object in its native format and converts it to an R object. Specifically, this function implements an algorithm to an import Excel workbook sheets. The function returns a tibble list (a list of tibbles).
}
