% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_initial_ls.R
\name{initial_ls}
\alias{initial_ls}
\alias{initial_ls-ready4fun_pkg_setup}
\alias{initial_ls,ready4fun_pkg_setup-method}
\title{initial_ls}
\usage{
initial_ls(x)

\S4method{initial_ls}{ready4fun_pkg_setup}(x)
}
\arguments{
\item{x}{An object of class ready4fun_pkg_setup}
}
\description{
S4 Generic function to get the value of the slot initial_ls

Get the value of the slot initial_ls for S4 objects of class ready4fun_pkg_setup
}
