% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_initial_ls.R
\name{initial_ls<-}
\alias{initial_ls<-}
\alias{initial_ls<--ready4fun_pkg_setup}
\alias{initial_ls<-,ready4fun_pkg_setup-method}
\title{initial_ls<-}
\usage{
initial_ls(x) <- value

\S4method{initial_ls}{ready4fun_pkg_setup}(x) <- value
}
\arguments{
\item{x}{An object of class ready4fun_pkg_setup}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot initial_ls

Set the value of the slot initial_ls for S4 objects of class ready4fun_pkg_setup
}
