% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_executor.R
\name{is_ready4fun_executor}
\alias{is_ready4fun_executor}
\title{is ready4fun package executor ready4 S3 class for list object specifying function arguments and function.}
\usage{
is_ready4fun_executor(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for list object specifying function arguments and function.
}
\description{
is ready4fun package executor ready4 S3 class for list object specifying function arguments and function.
}
