% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_executor.R
\name{is_ready4fun_executor}
\alias{is_ready4fun_executor}
\title{Is ready4fun package function list ready4 S3 class for declaring package description file data.}
\usage{
is_ready4fun_executor(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for declaring package description file data.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for declaring package description file data.
}
\details{
ready4 S3 class for list object specifying function arguments and function.
}
