% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_manifest.R
\name{is_ready4fun_manifest}
\alias{is_ready4fun_manifest}
\title{Is ready4fun package manifest ready4 S3 class for encapsulating the metadata required for package set-up.}
\usage{
is_ready4fun_manifest(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for encapsulating the metadata required for package set-up.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for encapsulating the metadata required for package set-up.
}
\details{
ready4 S3 class for encapsulating the metadata required for package set-up.
}
