% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_manifest_one.R
\name{is_ready4fun_manifest_one}
\alias{is_ready4fun_manifest_one}
\title{Is ready4fun package manifest one ready4 S3 class for package metadata required for initial package set-up step.}
\usage{
is_ready4fun_manifest_one(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for package metadata required for initial package set-up step.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for package metadata required for initial package set-up step.
}
\details{
ready4 S3 class for package metadata required for initial package set-up step.
}
