% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_metadata_b.R
\name{is_ready4fun_metadata_b}
\alias{is_ready4fun_metadata_b}
\title{is ready4fun package metadata b ready4 S3 class for package metadata required for second package set-up step.}
\usage{
is_ready4fun_metadata_b(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for package metadata required for second package set-up step.
}
\description{
is ready4fun package metadata b ready4 S3 class for package metadata required for second package set-up step.
}
