% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_description.R
\name{is_ready4fun_description}
\alias{is_ready4fun_description}
\title{Is ready4fun package package description ready4 S4 class for declaring package description file data.}
\usage{
is_ready4fun_description(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S4 class for declaring package description file data.
}
\description{
Check whether an object is a valid instance of the ready4 S4 class for declaring package description file data.
}
\details{
ready4 S3 class for declaring package description file data.
}
