% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_ds.R
\name{is_ready4fun_pkg_ds}
\alias{is_ready4fun_pkg_ds}
\title{Is ready4fun package dataset ready4 S4 class for declaring package description file data..1}
\usage{
is_ready4fun_pkg_ds(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S4 class for declaring package description file data..1
}
\description{
Check whether an object is a valid instance of the ready4 S4 class for declaring package description file data..1
}
\details{
ready4 S3 class for declaring package description file data.
}
