% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_setup.R
\name{is_ready4fun_pkg_setup}
\alias{is_ready4fun_pkg_setup}
\title{Is ready4 S4 class for package metadata required for package set-up.}
\usage{
is_ready4fun_pkg_setup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S4 class for package metadata required for package set-up.
}
\description{
Check whether an object is a valid instance of the ready4 S4 class for package metadata required for package set-up.
}
\details{
ready4 S3 class for package metadata required for package set-up.
}
