% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{make}
\alias{make}
\title{Make}
\usage{
make(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
make() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make.
}
