% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_addl_pkgs_ls}
\alias{make_addl_pkgs_ls}
\title{Make additional packages list}
\usage{
make_addl_pkgs_ls(
  depends_chr = NULL,
  enhances_chr = NULL,
  imports_chr = NULL,
  linking_to_chr = NULL,
  suggests_chr = NULL,
  append_ls = NULL
)
}
\arguments{
\item{depends_chr}{Depends (a character vector), Default: NULL}

\item{enhances_chr}{Enhances (a character vector), Default: NULL}

\item{imports_chr}{Imports (a character vector), Default: NULL}

\item{linking_to_chr}{Linking to (a character vector), Default: NULL}

\item{suggests_chr}{Suggests (a character vector), Default: NULL}

\item{append_ls}{Append (a list), Default: NULL}
}
\value{
Additional packages (a list)
}
\description{
make_addl_pkgs_ls() is a Make function that creates a new R object Specifically, this function implements an algorithm to make additional packages list. The function returns Additional packages (a list).
}
\keyword{internal}
