% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_all_fns_dmt_tb}
\alias{make_all_fns_dmt_tb}
\title{Make all functions dmt}
\usage{
make_all_fns_dmt_tb(
  paths_ls,
  undocumented_fns_dir_chr = make_undmtd_fns_dir_chr(),
  custom_dmt_ls = list(details_ls = NULL, export_ls = list(force_true_chr_vec =
    NA_character_, force_false_chr_vec = NA_character_), args_ls_ls = NULL),
  fn_type_lup_tb,
  abbreviations_lup = NULL
)
}
\arguments{
\item{paths_ls}{Paths (a list)}

\item{undocumented_fns_dir_chr}{Undocumented functions directory (a character vector of length 1), Default: make_undmtd_fns_dir_chr()}

\item{custom_dmt_ls}{Custom dmt (a list), Default: list(details_ls = NULL, export_ls = list(force_true_chr_vec = NA_character_, 
force_false_chr_vec = NA_character_), args_ls_ls = NULL)}

\item{fn_type_lup_tb}{Function type lookup table (a tibble)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\value{
All functions dmt (a tibble)
}
\description{
make_all_fns_dmt_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make all a functions dmt.The function returns all functions dmt (a tibble).
}
\keyword{internal}
