% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_and_doc_generics_tb_R}
\alias{make_and_doc_generics_tb_R}
\title{Make and document generics tibble}
\usage{
make_and_doc_generics_tb_R(
  generic_nm_chr,
  description_chr,
  overwrite_lgl = T,
  pkg_nm_chr,
  url_chr = NA_character_,
  abbreviations_lup = NULL
)
}
\arguments{
\item{generic_nm_chr}{Generic name (a character vector of length 1)}

\item{description_chr}{Description (a character vector of length 1)}

\item{overwrite_lgl}{Overwrite (a logical vector of length 1), Default: T}

\item{pkg_nm_chr}{Package name (a character vector of length 1)}

\item{url_chr}{Url (a character vector of length 1), Default: 'NA'}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\description{
make_and_doc_generics_tb_R() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make and a document generics a tibble R.NA
}
\keyword{internal}
