% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_desc_spine}
\alias{make_arg_desc_spine}
\title{Make argument description spine}
\usage{
make_arg_desc_spine(
  argument_nm_1L_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  master_object_type_lup = NULL,
  server_1L_chr = deprecated()
)
}
\arguments{
\item{argument_nm_1L_chr}{Argument name (a character vector of length one)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{master_object_type_lup}{Master object type (a lookup table), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Argument description spine (a character vector of length one)
}
\description{
make_arg_desc_spine() is a Make function that creates a new R object Specifically, this function implements an algorithm to make argument description spine. The function returns Argument description spine (a character vector of length one).
}
\keyword{internal}
