% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_desc_spine_chr}
\alias{make_arg_desc_spine_chr}
\title{Make argument description spine}
\usage{
make_arg_desc_spine_chr(
  argument_nm_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL
)
}
\arguments{
\item{argument_nm_chr}{Argument name (a character vector of length 1)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\value{
Argument description spine (a character vector of length 1)
}
\description{
make_arg_desc_spine_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make an argument description spine. The function returns an argument description spine (a character vector of length 1).
}
\keyword{internal}
