% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_title_chr_vec}
\alias{make_arg_title_chr_vec}
\title{Make argument title}
\usage{
make_arg_title_chr_vec(
  args_chr_vec,
  match_chr_vec,
  object_type_lup = NULL,
  abbreviations_lup = NULL
)
}
\arguments{
\item{args_chr_vec}{Arguments (a character vector)}

\item{match_chr_vec}{Match (a character vector)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\value{
Title (a character vector)
}
\description{
make_arg_title_chr_vec() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make an argument title. The function returns title (a character vector).
}
\keyword{internal}
