% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_type_abbr_spine_chr}
\alias{make_arg_type_abbr_spine_chr}
\title{Make argument type abbreviation spine}
\usage{
make_arg_type_abbr_spine_chr(argument_nm_chr, lup_tb)
}
\arguments{
\item{argument_nm_chr}{Argument name (a character vector of length 1)}

\item{lup_tb}{Lookup table (a tibble)}
}
\value{
Argument type abbreviation spine (a character vector of length 1)
}
\description{
make_arg_type_abbr_spine_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make an argument type an abbreviation spine.The function returns an argument type abbreviation spine (a character vector of length 1).
}
\keyword{internal}
