% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_type_lup_ls}
\alias{make_arg_type_lup_ls}
\title{Make argument type lookup table list}
\usage{
make_arg_type_lup_ls(
  object_type_lup = NULL,
  dv_ds_nm_1L_chr = "https://doi.org/10.7910/DVN/2Y9VF9",
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'https://doi.org/10.7910/DVN/2Y9VF9'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
Lookup table list (a list of lookup tables)
}
\description{
make_arg_type_lup_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make argument type lookup table list. The function returns Lookup table list (a list of lookup tables).
}
