% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_build_ignore_ls}
\alias{make_build_ignore_ls}
\title{Make build ignore list}
\usage{
make_build_ignore_ls(file_nms_chr = NULL, regulars_rgx = NULL)
}
\arguments{
\item{file_nms_chr}{File names (a character vector), Default: NULL}

\item{regulars_rgx}{Regulars (a regular expression vector), Default: NULL}
}
\value{
Build ignore (a list)
}
\description{
make_build_ignore_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make build ignore list. The function returns Build ignore (a list).
}
\keyword{internal}
