% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_custom_dmt_ls}
\alias{make_custom_dmt_ls}
\title{Make custom documentation list}
\usage{
make_custom_dmt_ls(
  args_ls_ls = NULL,
  desc_ls = NULL,
  details_ls = NULL,
  example_ls = NULL,
  output_ls = NULL,
  title_ls = NULL,
  user_manual_fns_chr = NA_character_
)
}
\arguments{
\item{args_ls_ls}{Arguments (a list of lists), Default: NULL}

\item{desc_ls}{Description (a list), Default: NULL}

\item{details_ls}{Details (a list), Default: NULL}

\item{example_ls}{Example (a list), Default: NULL}

\item{output_ls}{Output (a list), Default: NULL}

\item{title_ls}{Title (a list), Default: NULL}

\item{user_manual_fns_chr}{User manual functions (a character vector), Default: 'NA'}
}
\value{
Custom documentation (a list)
}
\description{
make_custom_dmt_ls() is a Make function that creates a new R object Specifically, this function implements an algorithm to make custom documentation list. The function returns Custom documentation (a list).
}
\keyword{internal}
