% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_depnt_fns_ls}
\alias{make_depnt_fns_ls}
\title{Make dependent functions list}
\usage{
make_depnt_fns_ls(arg_ls, pkg_depcy_ls)
}
\arguments{
\item{arg_ls}{Argument (a list)}

\item{pkg_depcy_ls}{Package dependency (a list)}
}
\value{
Argument (a list)
}
\description{
make_depnt_fns_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dependent functions list. The function returns Argument (a list).
}
