% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_desc}
\alias{make_fn_desc}
\title{Make function description}
\usage{
make_fn_desc(
  fns_chr,
  title_chr,
  output_chr,
  fn_type_lup_tb = NULL,
  abbreviations_lup = NULL
)
}
\arguments{
\item{fns_chr}{Functions (a character vector)}

\item{title_chr}{Title (a character vector)}

\item{output_chr}{Output (a character vector)}

\item{fn_type_lup_tb}{Function type lookup table (a tibble), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\value{
Function description (a character vector)
}
\description{
make_fn_desc() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a function description.The function returns a function description (a character vector).
}
\keyword{internal}
