% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_desc_chr_vec}
\alias{make_fn_desc_chr_vec}
\title{Make function description}
\usage{
make_fn_desc_chr_vec(fns_chr_vec, title_chr_vec, output_chr_vec)
}
\arguments{
\item{fns_chr_vec}{Functions (a character vector)}

\item{title_chr_vec}{Title (a character vector)}

\item{output_chr_vec}{Output (a character vector)}
}
\value{
Function description (a character vector)
}
\description{
make_fn_desc_chr_vec() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a function description. The function returns a function description (a character vector).
}
\keyword{internal}
