% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_desc_spine}
\alias{make_fn_desc_spine}
\title{Make function description spine}
\usage{
make_fn_desc_spine(
  fn_name_1L_chr,
  fn_title_1L_chr,
  fn_type_lup_tb = NULL,
  abbreviations_lup = NULL
)
}
\arguments{
\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{fn_title_1L_chr}{Function title (a character vector of length one)}

\item{fn_type_lup_tb}{Function type lookup table (a tibble), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\value{
Function description spine (a character vector of length one)
}
\description{
make_fn_desc_spine() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function description spine. The function returns a function description spine (a character vector of length one).
}
