% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_desc_spine_chr_vec}
\alias{make_fn_desc_spine_chr_vec}
\title{Make function description spine}
\usage{
make_fn_desc_spine_chr_vec(fn_name_chr, fn_title_chr, fn_type_lup_tb = NULL)
}
\arguments{
\item{fn_name_chr}{Function name (a character vector of length 1)}

\item{fn_title_chr}{Function title (a character vector of length 1)}

\item{fn_type_lup_tb}{Function type lookup table (a tibble), Default: NULL}
}
\value{
Function description spine (a character vector)
}
\description{
make_fn_desc_spine_chr_vec() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a function description spine. The function returns a function description spine (a character vector).
}
\keyword{internal}
