% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_dmt_spine}
\alias{make_fn_dmt_spine}
\title{Make function documentation spine}
\usage{
make_fn_dmt_spine(
  fn_name_1L_chr,
  fn_type_1L_chr,
  fn_title_1L_chr = NA_character_,
  fn,
  details_1L_chr = NA_character_,
  example_1L_lgl = F,
  export_1L_lgl = T,
  class_name_1L_chr,
  doc_in_class_1L_lgl
)
}
\arguments{
\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{fn_type_1L_chr}{Function type (a character vector of length one)}

\item{fn_title_1L_chr}{Function title (a character vector of length one), Default: 'NA'}

\item{fn}{Function (a function)}

\item{details_1L_chr}{Details (a character vector of length one), Default: 'NA'}

\item{example_1L_lgl}{Example (a logical vector of length one), Default: F}

\item{export_1L_lgl}{Export (a logical vector of length one), Default: T}

\item{class_name_1L_chr}{Class name (a character vector of length one)}

\item{doc_in_class_1L_lgl}{Document in class (a logical vector of length one)}
}
\value{
Function documentation spine (a list of character vectors)
}
\description{
make_fn_dmt_spine() is a Make function that creates a new R object. Specifically, this function implements an algorithm to Make function documentation spine.The function returns a function documentation spine (a list of character vectors).
}
\keyword{internal}
