% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_dmt_tbl_tpl}
\alias{make_fn_dmt_tbl_tpl}
\title{Make function documentation table template}
\usage{
make_fn_dmt_tbl_tpl(
  fns_path_chr,
  fns_dir_chr = make_undmtd_fns_dir_chr(),
  fn_type_lup_tb = NULL,
  abbreviations_lup = NULL,
  object_type_lup = NULL
)
}
\arguments{
\item{fns_path_chr}{Functions path (a character vector)}

\item{fns_dir_chr}{Functions directory (a character vector), Default: make_undmtd_fns_dir_chr()}

\item{fn_type_lup_tb}{Function type lookup table (a tibble), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}
}
\value{
Function documentation table (a tibble)
}
\description{
make_fn_dmt_tbl_tpl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a function documentation a table template.The function returns a function documentation table (a tibble).
}
\keyword{internal}
