% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_nms}
\alias{make_fn_nms}
\title{Make function names}
\usage{
make_fn_nms(path_1L_chr = "data-raw")
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one), Default: 'data-raw'}
}
\value{
Functions (a list of character vectors of length one)
}
\description{
make_fn_nms() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function names.The function returns a functions list of character vectors of length one.
}
