% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_title_chr}
\alias{make_fn_title_chr}
\title{Make function title}
\usage{
make_fn_title_chr(
  fns_chr_vec,
  object_type_lup = NULL,
  abbreviations_lup = NULL,
  is_generic_lgl = F
)
}
\arguments{
\item{fns_chr_vec}{Functions (a character vector)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{is_generic_lgl}{Is generic (a logical vector of length 1), Default: F}
}
\value{
Title (a character vector)
}
\description{
make_fn_title_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a function title.The function returns title (a character vector).
}
\keyword{internal}
