% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_type_lup_tb}
\alias{make_fn_type_lup_tb}
\title{Make function type lookup table}
\usage{
make_fn_type_lup_tb(
  fn_type_nm_chr = character(0),
  fn_type_desc_chr = character(0),
  first_arg_desc_chr = character(0),
  second_arg_desc_chr = character(0),
  is_generic_lgl = logical(0),
  is_method_lgl = logical(0)
)
}
\arguments{
\item{fn_type_nm_chr}{Function type name (a character vector of length 1), Default: character(0)}

\item{fn_type_desc_chr}{Function type description (a character vector of length 1), Default: character(0)}

\item{first_arg_desc_chr}{First argument description (a character vector of length 1), Default: character(0)}

\item{second_arg_desc_chr}{Second argument description (a character vector of length 1), Default: character(0)}

\item{is_generic_lgl}{Is generic (a logical vector of length 1), Default: logical(0)}

\item{is_method_lgl}{Is method (a logical vector of length 1), Default: logical(0)}
}
\value{
Function type lookup table (a tibble)
}
\description{
make_fn_type_lup_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a function type a lookup table.The function returns a function type lookup table (a tibble).
}
\keyword{internal}
