% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_types}
\alias{make_fn_types}
\title{Make function types}
\usage{
make_fn_types()
}
\value{
Functions type (a character vector)
}
\description{
make_fn_types() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function types. The function returns a functions type (a character vector).
}
