% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fns_chr_ls}
\alias{make_fns_chr_ls}
\title{Make functions}
\usage{
make_fns_chr_ls(path_1L_chr = "data-raw")
}
\arguments{
\item{path_1L_chr}{Path 1L (a character vector of length 1), Default: 'data-raw'}
}
\value{
Functions (a list of character vectors of length 1)
}
\description{
make_fns_chr_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a functions.The function returns a functions (a list of character vectors of length 1).
}
\keyword{internal}
