% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fns_type_chr}
\alias{make_fns_type_chr}
\title{Make functions type}
\usage{
make_fns_type_chr()
}
\value{
Functions type (a character vector of length 1)
}
\description{
make_fns_type_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a functions type.The function returns a functions type (a character vector of length 1).
}
\keyword{internal}
