% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gtr_str_dmt_spine}
\alias{make_gtr_str_dmt_spine}
\title{Make getter setter documentation spine}
\usage{
make_gtr_str_dmt_spine(
  fn_type_1L_chr,
  fn_name_1L_chr,
  class_name_1L_chr,
  doc_in_class_1L_lgl,
  example_1L_lgl = F
)
}
\arguments{
\item{fn_type_1L_chr}{Function type (a character vector of length one)}

\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{class_name_1L_chr}{Class name (a character vector of length one)}

\item{doc_in_class_1L_lgl}{Document in class (a logical vector of length one)}

\item{example_1L_lgl}{Example (a logical vector of length one), Default: F}
}
\value{
Getter setter documentation spine (a list of character vectors)
}
\description{
make_gtr_str_dmt_spine() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make getter setter documentation spine. The function returns Getter setter documentation spine (a list of character vectors).
}
\keyword{internal}
