% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gtr_str_dmt_spine_chr_ls}
\alias{make_gtr_str_dmt_spine_chr_ls}
\title{Make getter setter dmt spine}
\usage{
make_gtr_str_dmt_spine_chr_ls(
  fn_type_chr,
  fn_name_chr,
  class_name_chr,
  doc_in_class_lgl,
  example_lgl = F
)
}
\arguments{
\item{fn_type_chr}{Function type (a character vector of length 1)}

\item{fn_name_chr}{Function name (a character vector of length 1)}

\item{class_name_chr}{Class name (a character vector of length 1)}

\item{doc_in_class_lgl}{Document in class (a logical vector of length 1)}

\item{example_lgl}{Example (a logical vector of length 1), Default: F}
}
\value{
Getter setter dmt spine (a list of character vectors of length 1)
}
\description{
make_gtr_str_dmt_spine_chr_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a getter setter dmt spine. The function returns a getter setter dmt spine (a list of character vectors of length 1).
}
\keyword{internal}
