% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_new_entries_tb}
\alias{make_new_entries_tb}
\title{Make new entries tibble}
\usage{
make_new_entries_tb(
  short_name_chr,
  long_name_chr,
  atomic_element_lgl = F,
  r3_can_extend_lgl = F
)
}
\arguments{
\item{short_name_chr}{Short name (a character vector)}

\item{long_name_chr}{Long name (a character vector)}

\item{atomic_element_lgl}{Atomic element (a logical vector), Default: F}

\item{r3_can_extend_lgl}{Ready4 S3 can extend (a logical vector), Default: F}
}
\value{
New entries (a tibble)
}
\description{
make_new_entries_tb() is a Make function that creates a new R object Specifically, this function implements an algorithm to make new entries tibble. The function returns New entries (a tibble).
}
\keyword{internal}
