% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_badges.R
\name{make_new_ready4fun_badges}
\alias{make_new_ready4fun_badges}
\title{Make new ready4 S3 class for tibble object lookup table of badges metadata.}
\usage{
make_new_ready4fun_badges(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object lookup table of badges metadata.}
}
\value{
An unvalidated instance of the ready4 S3 class for tibble object lookup table of badges metadata.
}
\description{
Create a new unvalidated instance of the ready4 S3 class for tibble object lookup table of badges metadata.
}
\details{
ready4 S3 class for tibble object lookup table of badges metadata.
}
