% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_executor.R
\name{make_new_ready4fun_executor}
\alias{make_new_ready4fun_executor}
\title{Make new ready4fun package executor ready4 S3 class for list object specifying function arguments and function.}
\usage{
make_new_ready4fun_executor(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for list object specifying function arguments and function.}
}
\value{
An unvalidated instance of the ready4 S3 class for list object specifying function arguments and function.
}
\description{
Create a new unvalidated instance of the ready4 S3 class for list object specifying function arguments and function.
}
\details{
ready4 S3 class for list object specifying function arguments and function.
}
