% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_fn_ls.R
\name{make_new_ready4fun_fn_ls}
\alias{make_new_ready4fun_fn_ls}
\title{Make new ready4fun package function list ready4 S3 class for declaring package description file data.}
\usage{
make_new_ready4fun_fn_ls(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for declaring package description file data.}
}
\value{
An unvalidated instance of the ready4 S3 class for declaring package description file data.
}
\description{
Create a new unvalidated instance of the ready4 S3 class for declaring package description file data.
}
\details{
ready4 S3 class for list object specifying function arguments and function.
}
