% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_functions.R
\name{make_new_ready4fun_functions}
\alias{make_new_ready4fun_functions}
\title{make new ready4fun package functions Function types lookup table.}
\usage{
make_new_ready4fun_functions(x)
}
\arguments{
\item{x}{A prototype for the Function types lookup table.}
}
\value{
An unvalidated instance of the Function types lookup table.
}
\description{
Create a new unvalidated instance of the Function types lookup table.
}
\details{
Function types lookup table.
}
\keyword{internal}
