% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_metadata_b.R
\name{make_new_ready4fun_metadata_b}
\alias{make_new_ready4fun_metadata_b}
\title{make new ready4fun package metadata b ready4 S3 class for package metadata required for second package set-up step.}
\usage{
make_new_ready4fun_metadata_b(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for package metadata required for second package set-up step.}
}
\value{
An unvalidated instance of the ready4 S3 class for package metadata required for second package set-up step.
}
\description{
Create a new unvalidated instance of the ready4 S3 class for package metadata required for second package set-up step.
}
\details{
ready4 S3 class for package metadata required for second package set-up step.
}
\keyword{internal}
