% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_objects.R
\name{make_new_ready4fun_objects}
\alias{make_new_ready4fun_objects}
\title{make new ready4fun package objects Object types lookup table.}
\usage{
make_new_ready4fun_objects(x)
}
\arguments{
\item{x}{A prototype for the Object types lookup table.}
}
\value{
An unvalidated instance of the Object types lookup table.
}
\description{
Create a new unvalidated instance of the Object types lookup table.
}
\details{
Object types lookup table.
}
\keyword{internal}
