% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_desc.R
\name{make_new_ready4fun_pkg_desc}
\alias{make_new_ready4fun_pkg_desc}
\title{Make new ready4 S4 class for declaring package description file data.}
\usage{
make_new_ready4fun_pkg_desc(x)
}
\arguments{
\item{x}{A prototype for the ready4 S4 class for declaring package description file data.}
}
\value{
An unvalidated instance of the ready4 S4 class for declaring package description file data.
}
\description{
Create a new unvalidated instance of the ready4 S4 class for declaring package description file data.
}
\details{
ready4 S3 class for declaring package description file data.
}
