% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_dataset.R
\name{make_new_ready4fun_dataset}
\alias{make_new_ready4fun_dataset}
\title{Make new ready4fun package package dataset ready4 S4 class for declaring package description file data..1}
\usage{
make_new_ready4fun_dataset(x)
}
\arguments{
\item{x}{A prototype for the ready4 S4 class for declaring package description file data..1}
}
\value{
An unvalidated instance of the ready4 S4 class for declaring package description file data..1
}
\description{
Create a new unvalidated instance of the ready4 S4 class for declaring package description file data..1
}
\details{
ready4 S3 class for declaring package description file data.
}
