% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_setup.R
\name{make_new_ready4fun_pkg_setup}
\alias{make_new_ready4fun_pkg_setup}
\title{Make new ready4fun package package setup ready4 S4 class for package metadata required for package set-up.}
\usage{
make_new_ready4fun_pkg_setup(x)
}
\arguments{
\item{x}{A prototype for the ready4 S4 class for package metadata required for package set-up.}
}
\value{
An unvalidated instance of the ready4 S4 class for package metadata required for package set-up.
}
\description{
Create a new unvalidated instance of the ready4 S4 class for package metadata required for package set-up.
}
\details{
ready4 S3 class for package metadata required for package set-up.
}
