% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_obj_lup}
\alias{make_obj_lup}
\title{Make object}
\usage{
make_obj_lup()
}
\value{
Object (a tibble)
}
\description{
make_obj_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make object lookup table. The function returns an object (a tibble).
}
