% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_obj_lup_tb}
\alias{make_obj_lup_tb}
\title{Make object lookup table}
\usage{
make_obj_lup_tb()
}
\value{
Object (a tibble)
}
\description{
make_obj_lup_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make an object lookup table.The function returns an object (a tibble).
}
\keyword{internal}
