% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pkg_desc_ls}
\alias{make_pkg_desc_ls}
\title{Make package description list}
\usage{
make_pkg_desc_ls(
  pkg_nm_1L_chr = get_dev_pkg_nm(),
  pkg_title_1L_chr,
  pkg_desc_1L_chr,
  authors_prsn,
  urls_chr
)
}
\arguments{
\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: get_dev_pkg_nm()}

\item{pkg_title_1L_chr}{Package title (a character vector of length one)}

\item{pkg_desc_1L_chr}{Package description (a character vector of length one)}

\item{authors_prsn}{Authors (a person)}

\item{urls_chr}{Urls (a character vector)}
}
\value{
Package description (a list)
}
\description{
make_pkg_desc_ls() is a Make function that creates a new R object Specifically, this function implements an algorithm to make package description list. The function returns Package description (a list).
}
\keyword{internal}
