% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pkg_ds_ls}
\alias{make_pkg_ds_ls}
\title{Make package dataset list}
\usage{
make_pkg_ds_ls(
  db_df,
  db_1L_chr,
  title_1L_chr,
  desc_1L_chr,
  abbreviations_lup = NULL,
  format_1L_chr = "A tibble",
  object_type_lup = NULL,
  simple_lup_1L_lgl = F,
  url_1L_chr = NA_character_,
  vars_ls = NULL
)
}
\arguments{
\item{db_df}{Database (a data.frame)}

\item{db_1L_chr}{Database (a character vector of length one)}

\item{title_1L_chr}{Title (a character vector of length one)}

\item{desc_1L_chr}{Description (a character vector of length one)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{format_1L_chr}{Format (a character vector of length one), Default: 'A tibble'}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{simple_lup_1L_lgl}{Simple lookup table (a logical vector of length one), Default: F}

\item{url_1L_chr}{Url (a character vector of length one), Default: 'NA'}

\item{vars_ls}{Variables (a list), Default: NULL}
}
\value{
Package dataset (a list)
}
\description{
make_pkg_ds_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make package dataset list. The function returns Package dataset (a list).
}
