% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_abbreviations.R
\name{make_pt_ready4fun_abbreviations}
\alias{make_pt_ready4fun_abbreviations}
\title{Make prototype ready4fun abbreviations ready4 S3 class for tibble object lookup table of abbreviations.}
\usage{
make_pt_ready4fun_abbreviations(
  short_name_chr = character(0),
  long_name_chr = character(0),
  plural_lgl = logical(0)
)
}
\arguments{
\item{short_name_chr}{Short name (a character vector), Default: character(0)}

\item{long_name_chr}{Long name (a character vector), Default: character(0)}

\item{plural_lgl}{Plural (a logical vector), Default: logical(0)}
}
\value{
A prototype for ready4 S3 class for tibble object lookup table of abbreviations.
}
\description{
Create a new prototype for the ready4 S3 class for tibble object lookup table of abbreviations.
}
\details{
ready4 S3 class for tibble object lookup table of abbreviations.
}
