% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_badges.R
\name{make_pt_ready4fun_badges}
\alias{make_pt_ready4fun_badges}
\title{Make prototype ready4fun badges ready4 S3 class for tibble object lookup table of badges metadata.}
\usage{
make_pt_ready4fun_badges(
  badge_names_chr = character(0),
  label_names_chr = character(0),
  colours_chr = character(0),
  badges_chr = character(0)
)
}
\arguments{
\item{badge_names_chr}{Badge names (a character vector), Default: character(0)}

\item{label_names_chr}{Label names (a character vector), Default: character(0)}

\item{colours_chr}{Colours (a character vector), Default: character(0)}

\item{badges_chr}{Badges (a character vector), Default: character(0)}
}
\value{
A prototype for ready4 S3 class for tibble object lookup table of badges metadata.
}
\description{
Create a new prototype for the ready4 S3 class for tibble object lookup table of badges metadata.
}
\details{
ready4 S3 class for tibble object lookup table of badges metadata.
}
