% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_dataset.R
\name{make_pt_ready4fun_dataset}
\alias{make_pt_ready4fun_dataset}
\title{Make prototype ready4fun package dataset ready4 S3 class for declaring package description file data..1}
\usage{
make_pt_ready4fun_dataset(
  db_df = data.frame(),
  db_1L_chr = character(0),
  title_1L_chr = character(0),
  desc_1L_chr = character(0),
  abbreviations_lup = ready4fun_abbreviations(),
  format_1L_chr = character(0),
  object_type_lup = ready4fun_abbreviations(),
  simple_lup_1L_lgl = logical(0),
  url_1L_chr = character(0),
  vars_ls = list()
)
}
\arguments{
\item{db_df}{Database (a data.frame), Default: data.frame()}

\item{db_1L_chr}{Database (a character vector of length one), Default: character(0)}

\item{title_1L_chr}{Title (a character vector of length one), Default: character(0)}

\item{desc_1L_chr}{Description (a character vector of length one), Default: character(0)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: ready4fun_abbreviations()}

\item{format_1L_chr}{Format (a character vector of length one), Default: character(0)}

\item{object_type_lup}{Object type (a lookup table), Default: ready4fun_abbreviations()}

\item{simple_lup_1L_lgl}{Simple lookup table (a logical vector of length one), Default: logical(0)}

\item{url_1L_chr}{Url (a character vector of length one), Default: character(0)}

\item{vars_ls}{Variables (a list), Default: list()}
}
\value{
A prototype for ready4 S3 class for declaring package description file data..1
}
\description{
Create a new prototype for the ready4 S3 class for declaring package description file data..1
}
\details{
ready4 S3 class for declaring package description file data.
}
