% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_description.R
\name{make_pt_ready4fun_description}
\alias{make_pt_ready4fun_description}
\title{make prototype ready4fun package description ready4 S3 class for declaring package description file data.}
\usage{
make_pt_ready4fun_description(
  Package = character(0),
  Title = character(0),
  Description = character(0),
  License = logical(0),
  URL = character(0)
)
}
\arguments{
\item{Package}{PARAM_DESCRIPTION, Default: character(0)}

\item{Title}{PARAM_DESCRIPTION, Default: character(0)}

\item{Description}{PARAM_DESCRIPTION, Default: character(0)}

\item{License}{PARAM_DESCRIPTION, Default: logical(0)}

\item{URL}{PARAM_DESCRIPTION, Default: character(0)}
}
\value{
A prototype for ready4 S3 class for declaring package description file data.
}
\description{
Create a new prototype for the ready4 S3 class for declaring package description file data.
}
\details{
ready4 S3 class for declaring package description file data.
}
