% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_executor.R
\name{make_pt_ready4fun_executor}
\alias{make_pt_ready4fun_executor}
\title{make prototype ready4fun package executor ready4 S3 class for list object specifying function arguments and function.}
\usage{
make_pt_ready4fun_executor(args_ls = list(), fn = identity)
}
\arguments{
\item{args_ls}{Arguments (a list), Default: list()}

\item{fn}{Function (a function), Default: identity}
}
\value{
A prototype for ready4 S3 class for list object specifying function arguments and function.
}
\description{
make prototype ready4fun package executor ready4 S3 class for list object specifying function arguments and function.
}
