% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_manifest.R
\name{make_pt_ready4fun_manifest}
\alias{make_pt_ready4fun_manifest}
\title{Make prototype ready4fun package manifest ready4 S3 class for encapsulating the metadata required for package set-up.}
\usage{
make_pt_ready4fun_manifest(
  initial_ls = ready4fun_manifest_one(),
  subsequent_ls = ready4fun_manifest_two()
)
}
\arguments{
\item{initial_ls}{Initial (a list), Default: ready4fun_manifest_one()}

\item{subsequent_ls}{Subsequent (a list), Default: ready4fun_manifest_two()}
}
\value{
A prototype for ready4 S3 class for encapsulating the metadata required for package set-up.
}
\description{
Create a new prototype for the ready4 S3 class for encapsulating the metadata required for package set-up.
}
\details{
ready4 S3 class for encapsulating the metadata required for package set-up.
}
