% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_setup_one.R
\name{make_pt_ready4fun_pkg_setup_one}
\alias{make_pt_ready4fun_pkg_setup_one}
\title{Make prototype ready4fun package package setup one ready4 S4 class for package metadata required for initial package set-up step.}
\usage{
make_pt_ready4fun_pkg_setup_one(
  pkg_desc_ls = ready4fun_pkg_desc_ls(),
  copyright_holders_chr = character(0),
  gh_repo_1L_chr = character(0),
  add_gh_site_1L_lgl = character(0),
  addl_badges_ls = list(),
  badges_lup = ready4fun_badges(),
  check_type_1L_chr = character(0),
  delete_r_dir_cnts_1L_lgl = logical(0),
  dev_pkg_nm_1L_chr = character(0),
  lifecycle_stage_1L_chr = character(0),
  incr_ver_1L_lgl = logical(0),
  on_cran_1L_lgl = logical(0),
  path_to_pkg_logo_1L_chr = character(0),
  path_to_pkg_rt_1L_chr = character(0)
)
}
\arguments{
\item{pkg_desc_ls}{Package description (a list), Default: ready4fun_pkg_desc_ls()}

\item{copyright_holders_chr}{Copyright holders (a character vector), Default: character(0)}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: character(0)}

\item{add_gh_site_1L_lgl}{Add github site (a logical vector of length one), Default: character(0)}

\item{addl_badges_ls}{Additional badges (a list), Default: list()}

\item{badges_lup}{Badges (a lookup table), Default: ready4fun_badges()}

\item{check_type_1L_chr}{Check type (a character vector of length one), Default: character(0)}

\item{delete_r_dir_cnts_1L_lgl}{Delete r directory contents (a logical vector of length one), Default: logical(0)}

\item{dev_pkg_nm_1L_chr}{Development package name (a character vector of length one), Default: character(0)}

\item{lifecycle_stage_1L_chr}{Lifecycle stage (a character vector of length one), Default: character(0)}

\item{incr_ver_1L_lgl}{Increment version (a logical vector of length one), Default: logical(0)}

\item{on_cran_1L_lgl}{On cran (a logical vector of length one), Default: logical(0)}

\item{path_to_pkg_logo_1L_chr}{Path to package logo (a character vector of length one), Default: character(0)}

\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: character(0)}
}
\value{
A prototype for ready4 S4 class for package metadata required for initial package set-up step.
}
\description{
Create a new prototype for the ready4 S4 class for package metadata required for initial package set-up step.
}
\details{
ready4 S3 class for package metadata required for initial package set-up step.
}
