% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_setup_two.R
\name{make_pt_ready4fun_pkg_setup_two}
\alias{make_pt_ready4fun_pkg_setup_two}
\title{Make prototype ready4fun package setup two ready4 S4 class for package metadata required for second package set-up step.}
\usage{
make_pt_ready4fun_pkg_setup_two(
  addl_pkgs_ls = list(),
  build_ignore_ls = list(),
  dev_pkgs_chr = character(0),
  pkg_dmt_dv_dss_chr = character(0),
  user_manual_fns_chr = character(0)
)
}
\arguments{
\item{addl_pkgs_ls}{Additional packages (a list), Default: list()}

\item{build_ignore_ls}{Build ignore (a list), Default: list()}

\item{dev_pkgs_chr}{Development packages (a character vector), Default: character(0)}

\item{pkg_dmt_dv_dss_chr}{Package documentation dataverse datasets (a character vector), Default: character(0)}

\item{user_manual_fns_chr}{User manual functions (a character vector), Default: character(0)}
}
\value{
A prototype for ready4 S4 class for package metadata required for second package set-up step.
}
\description{
Create a new prototype for the ready4 S4 class for package metadata required for second package set-up step.
}
\details{
ready4 S3 class for package metadata required for second package set-up step.
}
