% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ret_obj_desc}
\alias{make_ret_obj_desc}
\title{Make ret object description}
\usage{
make_ret_obj_desc(fn, abbreviations_lup, starts_sentence_1L_lgl = T)
}
\arguments{
\item{fn}{Function (a function)}

\item{abbreviations_lup}{Abbreviations (a lookup table)}

\item{starts_sentence_1L_lgl}{Starts sentence (a logical vector of length one), Default: T}
}
\value{
Ret object description (a character vector of length one)
}
\description{
make_ret_obj_desc() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ret object description. The function returns Ret object description (a character vector of length one).
}
\keyword{internal}
