% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_undmtd_fns_dir_chr}
\alias{make_undmtd_fns_dir_chr}
\title{Make undmtd functions directory}
\usage{
make_undmtd_fns_dir_chr(path_1L_chr = "data-raw")
}
\arguments{
\item{path_1L_chr}{Path 1L (a character vector of length 1), Default: 'data-raw'}
}
\value{
Undocumented functions directory (a character vector of length 1)
}
\description{
make_undmtd_fns_dir_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make undmtd a functions directory.The function returns undocumented functions directory (a character vector of length 1).
}
\keyword{internal}
