% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_Package.R
\name{Package-ready4fun_pkg_desc}
\alias{Package-ready4fun_pkg_desc}
\alias{Package,ready4fun_pkg_desc-method}
\title{Package}
\usage{
\S4method{Package}{ready4fun_pkg_desc}(x)
}
\arguments{
\item{x}{An object of class ready4fun_pkg_desc}
}
\description{
Get the value of the slot Package for S4 objects of class ready4fun_pkg_desc
}
