% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_pkg_desc_ls.R
\name{pkg_desc_ls<--ready4fun_pkg_setup_one}
\alias{pkg_desc_ls<--ready4fun_pkg_setup_one}
\alias{pkg_desc_ls<-,ready4fun_pkg_setup_one-method}
\title{pkg_desc_ls<-}
\usage{
\S4method{pkg_desc_ls}{ready4fun_pkg_setup_one}(x) <- value
}
\arguments{
\item{x}{An object of class ready4fun_pkg_setup_one}
}
\description{
Set the value of the slot pkg_desc_ls for S4 objects of class ready4fun_pkg_setup_one
}
