% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure.ready4fun_abbreviations}
\alias{procure.ready4fun_abbreviations}
\alias{procure,ready4fun_abbreviations-method}
\alias{procure.ready4fun_manifest}
\alias{procure,ready4fun_manifest-method}
\title{Procure items from a dataset}
\usage{
procure.ready4fun_abbreviations(
  x,
  match_1L_chr = character(0),
  type_1L_chr = c("abbreviation", "extension"),
  what_1L_chr = c("table", "string")
)

\S4method{procure}{ready4fun_abbreviations}(
  x,
  match_1L_chr = character(0),
  type_1L_chr = c("abbreviation", "extension"),
  what_1L_chr = c("table", "string")
)

procure.ready4fun_manifest(x, type_1L_chr)

\S4method{procure}{ready4fun_manifest}(x, type_1L_chr)
}
\arguments{
\item{x}{An instance of \code{ready4fun_manifest}, a ready4 submodule class for encapsulating the metadata required for package set-up.}

\item{match_1L_chr}{Match (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("table", "string")}
}
\value{
Object (an output object of multiple potential types)

Value (an output object of multiple potential types)
}
\description{
procure.ready4fun_abbreviations() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of abbreviations. The function returns Object (an output object of multiple potential types).

procure.ready4fun_manifest() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for encapsulating the metadata required for package set-up. The function returns Value (an output object of multiple potential types).
}
