% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ratify.R
\name{ratify.ready4fun_manifest}
\alias{ratify.ready4fun_manifest}
\alias{ratify,ready4fun_manifest-method}
\title{Ratify method applied to ready4 S3 class for encapsulating the metadata required for package set-up..}
\usage{
ratify.ready4fun_manifest(x)

\S4method{ratify}{ready4fun_manifest}(x)
}
\arguments{
\item{x}{An instance of ready4 S3 class for encapsulating the metadata required for package set-up.}
}
\value{
X (ready4 S3 class for encapsulating the metadata required for package set-up.)
}
\description{
ratify.ready4fun_manifest() is a Ratify method that checks whether an instance of a class conforms to required criteria, potentially modifying that instance to ensure that it is valid. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
