% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_abbreviations.R
\name{ready4fun_abbreviations}
\alias{ready4fun_abbreviations}
\title{ready4 S3 class for tibble object lookup table of abbreviations.}
\usage{
ready4fun_abbreviations(x = make_pt_ready4fun_abbreviations())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object lookup table of abbreviations., Default: make_pt_ready4fun_abbreviations()}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table of abbreviations.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table of abbreviations.
}
\details{
ready4 S3 class for tibble object lookup table of abbreviations.
}
