% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_badges.R
\name{ready4fun_badges}
\alias{ready4fun_badges}
\title{ready4 S3 class for tibble object lookup table of badges metadata.}
\usage{
ready4fun_badges(x = make_pt_ready4fun_badges())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object lookup table of badges metadata., Default: make_pt_ready4fun_badges()}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table of badges metadata.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table of badges metadata.
}
\details{
ready4 S3 class for tibble object lookup table of badges metadata.
}
