% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_fn_ls.R
\name{ready4fun_fn_ls}
\alias{ready4fun_fn_ls}
\title{ready4 S3 class for declaring package description file data.}
\usage{
ready4fun_fn_ls(x = make_pt_ready4fun_fn_ls())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for declaring package description file data., Default: make_pt_ready4fun_fn_ls()}
}
\value{
A validated instance of the ready4 S3 class for declaring package description file data.
}
\description{
Create a new valid instance of the ready4 S3 class for declaring package description file data.
}
\details{
ready4 S3 class for declaring package description file data.
}
