% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_functions.R
\name{ready4fun_functions}
\alias{ready4fun_functions}
\alias{make_pt_ready4fun_functions}
\alias{is_ready4fun_functions}
\title{Function types lookup table.}
\usage{
ready4fun_functions(x = make_pt_ready4fun_functions())

make_pt_ready4fun_functions(
  fn_type_nm_chr = character(0),
  fn_type_desc_chr = character(0),
  first_arg_desc_chr = character(0),
  second_arg_desc_chr = character(0),
  is_generic_lgl = logical(0),
  is_method_lgl = logical(0)
)

is_ready4fun_functions(x)
}
\arguments{
\item{x}{An object of any type}

\item{fn_type_nm_chr}{Function type name (a character vector), Default: character(0)}

\item{fn_type_desc_chr}{Function type description (a character vector), Default: character(0)}

\item{first_arg_desc_chr}{First argument description (a character vector), Default: character(0)}

\item{second_arg_desc_chr}{Second argument description (a character vector), Default: character(0)}

\item{is_generic_lgl}{Is generic (a logical vector), Default: logical(0)}

\item{is_method_lgl}{Is method (a logical vector), Default: logical(0)}
}
\value{
A validated instance of the Function types lookup table.

A prototype for Function types lookup table.

A logical value, TRUE if a valid instance of the Function types lookup table.
}
\description{
Create a new valid instance of the Function types lookup table.
}
\details{
Function types lookup table.
}
