% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_manifest.R
\name{ready4fun_manifest}
\alias{ready4fun_manifest}
\title{ready4 S3 class for encapsulating the metadata required for package set-up.}
\usage{
ready4fun_manifest(x = make_pt_ready4fun_manifest())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for encapsulating the metadata required for package set-up., Default: make_pt_ready4fun_manifest()}
}
\value{
A validated instance of the ready4 S3 class for encapsulating the metadata required for package set-up.
}
\description{
Create a new valid instance of the ready4 S3 class for encapsulating the metadata required for package set-up.
}
\details{
ready4 S3 class for encapsulating the metadata required for package set-up.
}
