% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_manifest_two.R
\name{ready4fun_manifest_two}
\alias{ready4fun_manifest_two}
\title{ready4 S3 class for package metadata required for second package set-up step.}
\usage{
ready4fun_manifest_two(x = make_pt_ready4fun_manifest_two())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for package metadata required for second package set-up step., Default: make_pt_ready4fun_manifest_two()}
}
\value{
A validated instance of the ready4 S3 class for package metadata required for second package set-up step.
}
\description{
Create a new valid instance of the ready4 S3 class for package metadata required for second package set-up step.
}
\details{
ready4 S3 class for package metadata required for second package set-up step.
}
