% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_metadata_a.R
\name{ready4fun_metadata_a}
\alias{ready4fun_metadata_a}
\title{ready4 S3 class for package metadata required for initial package set-up step.}
\usage{
ready4fun_metadata_a(x = make_pt_ready4fun_metadata_a())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for package metadata required for initial package set-up step., Default: make_pt_ready4fun_metadata_a()}
}
\value{
A validated instance of the ready4 S3 class for package metadata required for initial package set-up step.
}
\description{
Create a new valid instance of the ready4 S3 class for package metadata required for initial package set-up step.
}
\details{
ready4 S3 class for package metadata required for initial package set-up step.
}
