% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_metadata_b.R
\name{ready4fun_metadata_b}
\alias{ready4fun_metadata_b}
\alias{make_pt_ready4fun_metadata_b}
\alias{is_ready4fun_metadata_b}
\title{for package metadata required for second package set-up step.}
\usage{
ready4fun_metadata_b(x = make_pt_ready4fun_metadata_b())

make_pt_ready4fun_metadata_b(
  abbreviations_lup = ready4fun_abbreviations(),
  addl_pkgs_ls = list(),
  build_ignore_ls = list(),
  cls_fn_ls = ready4fun_executor(),
  custom_dmt_ls = list(),
  dev_pkgs_chr = character(0),
  dss_records_ls = list(),
  dv_ds_nm_1L_chr = character(0),
  dv_url_pfx_1L_chr = character(0),
  fn_types_lup = tibble::tibble(),
  fns_dmt_tb = tibble::tibble(),
  import_from_chr = character(0),
  inc_pkg_meta_data_1L_lgl = logical(0),
  object_type_lup = ready4fun_abbreviations(),
  path_to_dmt_dir_1L_chr = character(0),
  piggyback_to_1L_chr = character(0),
  pkg_dmt_dv_dss_chr = character(0),
  pkg_ds_ls_ls = list(),
  seed_obj_type_lup = tibble::tibble(),
  server_1L_chr = character(0),
  s4_fns_ls = ready4fun_executor(),
  treat_as_words_chr = character(0)
)

is_ready4fun_metadata_b(x)
}
\arguments{
\item{x}{An object of any type}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: ready4fun_abbreviations()}

\item{addl_pkgs_ls}{Additional packages (a list), Default: list()}

\item{build_ignore_ls}{Build ignore (a list), Default: list()}

\item{cls_fn_ls}{Class (a list of functions), Default: ready4fun_executor()}

\item{custom_dmt_ls}{Custom documentation (a list), Default: list()}

\item{dev_pkgs_chr}{Development packages (a character vector), Default: character(0)}

\item{dss_records_ls}{Datasets records (a list), Default: list()}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: character(0)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: character(0)}

\item{fn_types_lup}{Function types (a lookup table), Default: tibble::tibble()}

\item{fns_dmt_tb}{Functions documentation (a tibble), Default: tibble::tibble()}

\item{import_from_chr}{Import from (a character vector), Default: character(0)}

\item{inc_pkg_meta_data_1L_lgl}{Include package meta data (a logical vector of length one), Default: logical(0)}

\item{object_type_lup}{Object type (a lookup table), Default: ready4fun_abbreviations()}

\item{path_to_dmt_dir_1L_chr}{Path to documentation directory (a character vector of length one), Default: character(0)}

\item{piggyback_to_1L_chr}{Piggyback to (a character vector of length one), Default: character(0)}

\item{pkg_dmt_dv_dss_chr}{Package documentation dataverse datasets (a character vector), Default: character(0)}

\item{pkg_ds_ls_ls}{Package dataset (a list of lists), Default: list()}

\item{seed_obj_type_lup}{Seed object type (a lookup table), Default: tibble::tibble()}

\item{server_1L_chr}{Server (a character vector of length one), Default: character(0)}

\item{s4_fns_ls}{S4 functions (a list), Default: ready4fun_executor()}

\item{treat_as_words_chr}{Treat as words (a character vector), Default: character(0)}
}
\value{
A validated instance of the ready4 S3 class for package metadata required for second package set-up step.

A prototype for ready4 S3 class for package metadata required for second package set-up step.

A logical value, TRUE if a valid instance of the ready4 S3 class for package metadata required for second package set-up step.
}
\description{
Create a new valid instance of the ready4 S3 class for package metadata required for second package set-up step.
}
\details{
ready4 S3 class for package metadata required for second package set-up step.
}
