% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_objects.R
\name{ready4fun_objects}
\alias{ready4fun_objects}
\alias{make_pt_ready4fun_objects}
\alias{is_ready4fun_objects}
\title{Object types lookup table.}
\usage{
ready4fun_objects(x = make_pt_ready4fun_objects())

make_pt_ready4fun_objects(
  short_name_chr = character(0),
  long_name_chr = character(0),
  atomic_element_lgl = logical(0),
  r3_can_extend_lgl = logical(0)
)

is_ready4fun_objects(x)
}
\arguments{
\item{x}{An object of any type}

\item{short_name_chr}{Short name (a character vector), Default: character(0)}

\item{long_name_chr}{Long name (a character vector), Default: character(0)}

\item{atomic_element_lgl}{Atomic element (a logical vector), Default: logical(0)}

\item{r3_can_extend_lgl}{Ready4 submodule can extend (a logical vector), Default: logical(0)}
}
\value{
A validated instance of the Object types lookup table.

A prototype for Object types lookup table.

A logical value, TRUE if a valid instance of the Object types lookup table.
}
\description{
Create a new valid instance of the Object types lookup table.
}
\details{
Object types lookup table.
}
