% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_setup.R
\name{ready4fun_pkg_setup}
\alias{ready4fun_pkg_setup}
\title{ready4 S4 class for package metadata required for package set-up.}
\usage{
ready4fun_pkg_setup(x = make_pt_ready4fun_pkg_setup())
}
\arguments{
\item{x}{A prototype for the ready4 S4 class for package metadata required for package set-up., Default: make_pt_ready4fun_pkg_setup()}
}
\value{
A validated instance of the ready4 S4 class for package metadata required for package set-up.
}
\description{
Create a new valid instance of the ready4 S4 class for package metadata required for package set-up.
}
\details{
ready4 S3 class for package metadata required for package set-up.
}
