% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_collate_chr}
\alias{remove_collate_chr}
\title{Remove collate}
\usage{
remove_collate_chr(description_chr)
}
\arguments{
\item{description_chr}{Description (a character vector of length 1)}
}
\value{
Description (a character vector of length 1)
}
\description{
remove_collate_chr() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove collate. Function argument description_chr specifies the object to be updated. Argument NA provides the object to be updated.The function returns a description (a character vector of length 1).
}
\keyword{internal}
