% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_obj_type_from_nm}
\alias{remove_obj_type_from_nm}
\title{Remove object type from name}
\usage{
remove_obj_type_from_nm(
  nms_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  fn_types_lup = NULL,
  is_generic_lgl = F,
  key_1L_chr = deprecated(),
  server_1L_chr = deprecated()
)
}
\arguments{
\item{nms_chr}{Names (a character vector)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{fn_types_lup}{Function types (a lookup table), Default: NULL}

\item{is_generic_lgl}{Is generic (a logical vector), Default: F}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Names (a character vector)
}
\description{
remove_obj_type_from_nm() is a Remove function that edits an object, removing a specified element or elements Specifically, this function implements an algorithm to remove object type from name. The function returns Names (a character vector).
}
\keyword{internal}
