% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_renew.R
\name{renew}
\alias{renew}
\alias{renew.ready4fun_manifest}
\alias{renew,ready4fun_manifest-method}
\title{Renew}
\usage{
renew(x, ...)

renew.ready4fun_manifest(
  x,
  type_1L_chr,
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)

\S4method{renew}{ready4fun_manifest}(
  x,
  type_1L_chr,
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{x}{An instance of ready4 S3 class for encapsulating the metadata required for package set-up.}

\item{...}{Additional arguments (an additional arguments)}

\item{type_1L_chr}{Type (a character vector of length one)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
X (ready4 S3 class for encapsulating the metadata required for package set-up.)
}
\description{
renew() is a Renew function that updates an instance of a class with new values. Specifically, this function implements an algorithm to renew. The function is called for its side effects and does not return a value.

renew.ready4fun_manifest() is a Renew method that updates an instance of a class with new values. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
