% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_replace.R
\name{replace_abbr_chr}
\alias{replace_abbr_chr}
\title{Replace abbreviation}
\usage{
replace_abbr_chr(title_chr, abbreviations_lup = NULL, collapse_lgl = T)
}
\arguments{
\item{title_chr}{Title (a character vector of length 1)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{collapse_lgl}{Collapse (a logical vector of length 1), Default: T}
}
\value{
Title (a character vector of length 1)
}
\description{
replace_abbr_chr() is a Replace function that edits an object, replacing a specified element with another specified element. Specifically, this function implements an algorithm to replace an abbreviation. Function argument title_chr specifies the object to be updated. Argument abbreviations_lup provides the object to be updated.The function returns title (a character vector of length 1).
}
\keyword{internal}
