% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_replace.R
\name{replace_fn_nms}
\alias{replace_fn_nms}
\title{Replace function names}
\usage{
replace_fn_nms(
  rename_tb,
  undocumented_fns_dir_chr = make_undmtd_fns_dir_chr(),
  rt_dev_dir_path_1L_chr = normalizePath("../../../"),
  dev_pkg_nm_1L_chr = get_dev_pkg_nm()
)
}
\arguments{
\item{rename_tb}{Rename (a tibble)}

\item{undocumented_fns_dir_chr}{Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr()}

\item{rt_dev_dir_path_1L_chr}{Root development directory path (a character vector of length one), Default: normalizePath("../../../")}

\item{dev_pkg_nm_1L_chr}{Development package name (a character vector of length one), Default: get_dev_pkg_nm()}
}
\description{
replace_fn_nms() is a Replace function that edits an object, replacing a specified element with another specified element. Specifically, this function implements an algorithm to a replace function names. Function argument rename_tb specifies the object to be updated. Argument undocumented_fns_dir_chr provides the object to be updated.The function is called for its side effects and does not return a value.
}
\keyword{internal}
