% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_report.R
\name{report.ready4fun_manifest}
\alias{report.ready4fun_manifest}
\alias{report,ready4fun_manifest-method}
\title{Report method applied to ready4 S3 class for encapsulating the metadata required for package set-up..}
\usage{
report.ready4fun_manifest(x, dv_url_pfx_1L_chr = NULL, key_1L_chr = NULL)

\S4method{report}{ready4fun_manifest}(x, dv_url_pfx_1L_chr = NULL, key_1L_chr = NULL)
}
\arguments{
\item{x}{An instance of ready4 S3 class for encapsulating the metadata required for package set-up.}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\description{
report.ready4fun_manifest() is a Report method that authors a report. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
