% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reset.R
\name{reset_pkg_files_R}
\alias{reset_pkg_files_R}
\title{Reset package files}
\usage{
reset_pkg_files_R(
  package_chr,
  package_dir = getwd(),
  description_ls = NULL,
  keep_version_lgl = T
)
}
\arguments{
\item{package_chr}{Package (a character vector of length 1)}

\item{package_dir}{PARAM_DESCRIPTION, Default: getwd()}

\item{description_ls}{Description (a list), Default: NULL}

\item{keep_version_lgl}{Keep version (a logical vector of length 1), Default: T}
}
\description{
reset_pkg_files_R() is a Reset function that edits an object, overwriting the current version with a default version. Specifically, this function implements an algorithm to reset a package files R.NA
}
\keyword{internal}
