% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_unload.R
\name{unload_packages}
\alias{unload_packages}
\title{Unload packages}
\usage{
unload_packages(package_chr)
}
\arguments{
\item{package_chr}{Package (a character vector)}
}
\description{
unload_packages() is an Unload function that performs a custom detaching of a package from the search path. Specifically, this function implements an algorithm to unload packages. Function argument package_chr specifies the package(s) to be detached from the search path. Argument NA provides the package(s) to be detached from the search path.NA
}
\keyword{internal}
