% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{update}
\alias{update}
\title{Update}
\usage{
update(object, ...)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
update() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update. Function argument object specifies the object to be updated. Argument ... provides the object to be updated.
}
