% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_abbr_lup}
\alias{update_abbr_lup}
\title{Update abbreviation}
\usage{
update_abbr_lup(
  abbr_tb,
  short_name_chr,
  long_name_chr,
  no_plural_chr = NA_character_,
  custom_plural_ls = NULL,
  pfx_rgx = NA_character_
)
}
\arguments{
\item{abbr_tb}{Abbreviation (a tibble)}

\item{short_name_chr}{Short name (a character vector)}

\item{long_name_chr}{Long name (a character vector)}

\item{no_plural_chr}{No plural (a character vector), Default: 'NA'}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}

\item{pfx_rgx}{Prefix (a regular expression vector), Default: 'NA'}
}
\value{
Abbreviation (a tibble)
}
\description{
update_abbr_lup() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update abbreviation lookup table. Function argument abbr_tb specifies the object to be updated. Argument short_name_chr provides the object to be updated. The function returns an abbreviation (a tibble).
}
\keyword{internal}
