% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_abbr_lup_tb}
\alias{update_abbr_lup_tb}
\title{Update abbreviation lookup table}
\usage{
update_abbr_lup_tb(
  abbr_tb,
  short_name_chr_vec,
  long_name_chr_vec,
  no_plural_chr_vec = NA_character_,
  custom_plural_ls = NULL,
  pfx_rgx_chr = NA_character_
)
}
\arguments{
\item{abbr_tb}{Abbreviation (a tibble)}

\item{short_name_chr_vec}{Short name (a character vector)}

\item{long_name_chr_vec}{Long name (a character vector)}

\item{no_plural_chr_vec}{No plural (a character vector), Default: 'NA'}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}

\item{pfx_rgx_chr}{Prefix regular expression (a character vector of length 1), Default: 'NA'}
}
\value{
Abbreviation (a tibble)
}
\description{
update_abbr_lup_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to an update abbreviation lookup table. Function argument abbr_tb specifies the object to be updated. Argument short_name_chr_vec provides the object to be updated.The function returns an abbreviation (a tibble).
}
\keyword{internal}
