% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_abbrs}
\alias{update_abbrs}
\title{Update abbreviations}
\usage{
update_abbrs(
  pkg_setup_ls,
  short_name_chr,
  long_name_chr,
  no_plural_chr = NA_character_,
  custom_plural_ls = NULL,
  pfx_rgx = NA_character_
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{short_name_chr}{Short name (a character vector)}

\item{long_name_chr}{Long name (a character vector)}

\item{no_plural_chr}{No plural (a character vector), Default: 'NA'}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}

\item{pfx_rgx}{Prefix (a regular expression vector), Default: 'NA'}
}
\value{
Package setup (a list)
}
\description{
update_abbrs() is an Update function that edits an object, while preserving core object attributes Specifically, this function implements an algorithm to update abbreviations. The function returns Package setup (a list).
}
\keyword{internal}
