% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_first_word_case}
\alias{update_first_word_case}
\title{Update first word case}
\usage{
update_first_word_case(phrase_1L_chr, fn = tolower)
}
\arguments{
\item{phrase_1L_chr}{Phrase (a character vector of length one)}

\item{fn}{Function (a function), Default: tolower}
}
\value{
Phrase (a character vector of length one)
}
\description{
update_first_word_case() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update first word case. Function argument phrase_1L_chr specifies the object to be updated. Argument fn provides the object to be updated. The function returns Phrase (a character vector of length one).
}
\keyword{internal}
