% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_first_word_case_chr}
\alias{update_first_word_case_chr}
\title{Update first word case}
\usage{
update_first_word_case_chr(phrase_chr, fn = tolower)
}
\arguments{
\item{phrase_chr}{Phrase (a character vector of length 1)}

\item{fn}{Function (a function), Default: tolower}
}
\value{
Phrase (a character vector of length 1)
}
\description{
update_first_word_case_chr() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to an update first word case. Function argument phrase_chr specifies the object to be updated. Argument fn provides the object to be updated. The function returns a phrase (a character vector of length 1).
}
\keyword{internal}
