% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_fn_dmt}
\alias{update_fn_dmt}
\title{Update function documentation}
\usage{
update_fn_dmt(
  fn_tags_spine_ls,
  new_tag_chr_ls,
  fn_name_1L_chr,
  fn_type_1L_chr,
  import_chr,
  import_from_chr = NA_character_,
  abbreviations_lup
)
}
\arguments{
\item{fn_tags_spine_ls}{Function tags spine (a list)}

\item{new_tag_chr_ls}{New tag (a list of character vectors)}

\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{fn_type_1L_chr}{Function type (a character vector of length one)}

\item{import_chr}{Import (a character vector)}

\item{import_from_chr}{Import from (a character vector), Default: 'NA'}

\item{abbreviations_lup}{Abbreviations (a lookup table)}
}
\value{
Function documentation (a character vector of length one)
}
\description{
update_fn_dmt() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update function documentation. Function argument fn_tags_spine_ls specifies the object to be updated. Argument new_tag_chr_ls provides the object to be updated. The function returns Function documentation (a character vector of length one).
}
\keyword{internal}
