% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_fn_dmt_chr}
\alias{update_fn_dmt_chr}
\title{Update function dmt}
\usage{
update_fn_dmt_chr(
  fn_tags_spine_ls,
  new_tag_chr_ls,
  fn_name_chr,
  fn_type_chr,
  import_chr_vec
)
}
\arguments{
\item{fn_tags_spine_ls}{Function tags spine (a list)}

\item{new_tag_chr_ls}{New tag (a list of character vectors of length 1)}

\item{fn_name_chr}{Function name (a character vector of length 1)}

\item{fn_type_chr}{Function type (a character vector of length 1)}

\item{import_chr_vec}{Import (a character vector)}
}
\value{
Function dmt (a character vector of length 1)
}
\description{
update_fn_dmt_chr() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to an update function dmt. Function argument fn_tags_spine_ls specifies the object to be updated. Argument new_tag_chr_ls provides the object to be updated. The function returns a function dmt (a character vector of length 1).
}
\keyword{internal}
