% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_fns_dmt_tb}
\alias{update_fns_dmt_tb}
\title{Update functions documentation}
\usage{
update_fns_dmt_tb(
  fns_dmt_tb,
  title_ls = NULL,
  desc_ls = NULL,
  details_ls = NULL,
  inc_for_main_user_lgl_ls = NULL,
  output_ls = NULL,
  example_ls = NULL,
  args_ls_ls = NULL,
  append_1L_lgl = T
)
}
\arguments{
\item{fns_dmt_tb}{Functions documentation (a tibble)}

\item{title_ls}{Title (a list), Default: NULL}

\item{desc_ls}{Description (a list), Default: NULL}

\item{details_ls}{Details (a list), Default: NULL}

\item{inc_for_main_user_lgl_ls}{Inc for main user (a list of logical vectors), Default: NULL}

\item{output_ls}{Output (a list), Default: NULL}

\item{example_ls}{Example (a list), Default: NULL}

\item{args_ls_ls}{Arguments (a list of lists), Default: NULL}

\item{append_1L_lgl}{Append (a logical vector of length one), Default: T}
}
\value{
Functions documentation (a tibble)
}
\description{
update_fns_dmt_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update functions documentation tibble. Function argument fns_dmt_tb specifies the object to be updated. Argument title_ls provides the object to be updated.The function returns a functions documentation (a tibble).
}
\keyword{internal}
