% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_fns_dmt_tb_lgl_vars}
\alias{update_fns_dmt_tb_lgl_vars}
\title{Update functions documentation tibble logical vector vars}
\usage{
update_fns_dmt_tb_lgl_vars(fns_dmt_tb, data_1L_chr, new_ls)
}
\arguments{
\item{fns_dmt_tb}{Functions documentation (a tibble)}

\item{data_1L_chr}{Data (a character vector of length one)}

\item{new_ls}{New (a list)}
}
\value{
Functions documentation (a tibble)
}
\description{
update_fns_dmt_tb_lgl_vars() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update functions documentation tibble logical vector vars. Function argument fns_dmt_tb specifies the object to be updated. Argument data_1L_chr provides the object to be updated.The function returns a functions documentation (a tibble).
}
\keyword{internal}
