% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_fns_dmt_tb_lgl_vars_chr_vec}
\alias{update_fns_dmt_tb_lgl_vars_chr_vec}
\title{Update functions documentation tibble logical vector of length 1 vars}
\usage{
update_fns_dmt_tb_lgl_vars_chr_vec(fns_dmt_tb, data_chr, new_ls)
}
\arguments{
\item{fns_dmt_tb}{Functions documentation (a tibble)}

\item{data_chr}{Data (a character vector of length 1)}

\item{new_ls}{New (a list)}
}
\value{
Functions documentation (a tibble)
}
\description{
update_fns_dmt_tb_lgl_vars_chr_vec() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to an update functions documentation a tibble logical vector of length 1 vars. Function argument fns_dmt_tb specifies the object to be updated. Argument data_chr provides the object to be updated. The function returns a functions documentation (a tibble).
}
\keyword{internal}
