% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_fns_dmt_tb_ls_vars_chr_vec}
\alias{update_fns_dmt_tb_ls_vars_chr_vec}
\title{Update functions dmt tibble list vars}
\usage{
update_fns_dmt_tb_ls_vars_chr_vec(fns_dmt_tb, data_chr, new_ls, append_lgl)
}
\arguments{
\item{fns_dmt_tb}{Functions dmt (a tibble)}

\item{data_chr}{Data (a character vector of length 1)}

\item{new_ls}{New (a list)}

\item{append_lgl}{Append (a logical vector of length 1)}
}
\value{
Functions dmt (a tibble)
}
\description{
update_fns_dmt_tb_ls_vars_chr_vec() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to an update functions dmt a tibble list vars. Function argument fns_dmt_tb specifies the object to be updated. Argument data_chr provides the object to be updated. The function returns a functions dmt (a tibble).
}
\keyword{internal}
