% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_msng_abbrs}
\alias{update_msng_abbrs}
\title{update missing abbreviations}
\usage{
update_msng_abbrs(
  pkg_setup_ls,
  are_words_chr = NA_character_,
  tf_to_singular_chr = NA_character_,
  not_obj_type_chr = NA_character_
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{are_words_chr}{Are words (a character vector), Default: 'NA'}

\item{tf_to_singular_chr}{Transform to singular (a character vector), Default: 'NA'}

\item{not_obj_type_chr}{Not object type (a character vector), Default: 'NA'}
}
\value{
Package setup (a list)
}
\description{
update_msng_abbrs() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update missing abbreviations. Function argument pkg_setup_ls specifies the object to be updated. Argument are_words_chr provides the object to be updated. The function returns Package setup (a list).
}
