% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_ns}
\alias{update_ns}
\title{Update namespace}
\usage{
update_ns(package_1L_chr)
}
\arguments{
\item{package_1L_chr}{Package (a character vector of length one)}
}
\value{
Package name (a character vector)
}
\description{
update_ns() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to Update namespace. Function argument package_1L_chr specifies the object to be updated.The function returns a package name (a character vector).
}
\keyword{internal}
