% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_ns_chr}
\alias{update_ns_chr}
\title{Update namespace}
\usage{
update_ns_chr(package_chr)
}
\arguments{
\item{package_chr}{Package (a character vector of length 1)}
}
\value{
Package name (a character vector of length 1)
}
\description{
update_ns_chr() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to an update namespace. Function argument package_chr specifies the object to be updated. Argument NA provides the object to be updated. The function returns a package name (a character vector of length 1).
}
