% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_pkg_setup_msgs}
\alias{update_pkg_setup_msgs}
\title{Update package setup messages}
\usage{
update_pkg_setup_msgs(pkg_setup_ls, list_element_1L_chr)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{list_element_1L_chr}{List element (a character vector of length one)}
}
\value{
Package setup (a list)
}
\description{
update_pkg_setup_msgs() is an Update function that edits an object, while preserving core object attributes Specifically, this function implements an algorithm to update package setup messages. The function returns Package setup (a list).
}
\keyword{internal}
