% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{validate}
\alias{validate}
\title{Validate}
\usage{
validate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
validate() is a Validate generic that validates that an object conforms to required criteria.
}
