% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ratify.R
\name{ratify.ready4fun_manifest}
\alias{ratify.ready4fun_manifest}
\title{Validate method applied to ready4 S3 class for encapsulating the metadata required for package set-up..}
\usage{
ratify.ready4fun_manifest(x)
}
\arguments{
\item{x}{An object}
}
\value{
X (ready4 S3 class for encapsulating the metadata required for package set-up.)
}
\description{
ratify.ready4fun_manifest() is a Validate method that validates that an object conforms to required criteria. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
