% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_validate.R
\name{validate_pkg_setup}
\alias{validate_pkg_setup}
\title{Validate package setup}
\usage{
validate_pkg_setup(
  pkg_setup_ls,
  classes_to_make_tb = NULL,
  pkg_ds_ls_ls = NULL
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{classes_to_make_tb}{Classes to make (a tibble), Default: NULL}

\item{pkg_ds_ls_ls}{Package dataset (a list of lists), Default: NULL}
}
\value{
Package setup (a list)
}
\description{
validate_pkg_setup() is a Validate function that validates that an object conforms to required criteria. Specifically, this function implements an algorithm to validate package setup. The function returns Package setup (a list).
}
\keyword{internal}
