% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_abbreviations.R
\name{validate_ready4fun_abbreviations}
\alias{validate_ready4fun_abbreviations}
\title{Validate ready4fun abbreviations ready4 S3 class for tibble object lookup table of abbreviations.}
\usage{
validate_ready4fun_abbreviations(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for tibble object lookup table of abbreviations.}
}
\value{
A prototpe for ready4 S3 class for tibble object lookup table of abbreviations.
}
\description{
Validate an instance of the ready4 S3 class for tibble object lookup table of abbreviations.
}
\details{
ready4 S3 class for tibble object lookup table of abbreviations.
}
