% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_badges.R
\name{validate_ready4fun_badges}
\alias{validate_ready4fun_badges}
\title{validate ready4fun package badges ready4 S3 class for tibble object lookup table of badges metadata.}
\usage{
validate_ready4fun_badges(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for tibble object lookup table of badges metadata.}
}
\value{
A prototpe for ready4 S3 class for tibble object lookup table of badges metadata.
}
\description{
Validate an instance of the ready4 S3 class for tibble object lookup table of badges metadata.
}
\details{
ready4 S3 class for tibble object lookup table of badges metadata.
}
\keyword{internal}
