% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_dataset.R
\name{validate_ready4fun_dataset}
\alias{validate_ready4fun_dataset}
\title{validate ready4fun package dataset ready4 S3 class for declaring package description file data..1}
\usage{
validate_ready4fun_dataset(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for declaring package description file data..1}
}
\value{
A prototpe for ready4 S3 class for declaring package description file data..1
}
\description{
Validate an instance of the ready4 S3 class for declaring package description file data..1
}
\details{
Package datasets metadata.
}
\keyword{internal}
