% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_executor.R
\name{validate_ready4fun_executor}
\alias{validate_ready4fun_executor}
\title{validate ready4fun package executor ready4 S3 class for list object specifying function arguments and function.}
\usage{
validate_ready4fun_executor(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for list object specifying function arguments and function.}
}
\value{
A prototpe for ready4 S3 class for list object specifying function arguments and function.
}
\description{
Validate an instance of the ready4 S3 class for list object specifying function arguments and function.
}
\details{
Function and function arguments pair.
}
\keyword{internal}
