% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_functions.R
\name{validate_ready4fun_functions}
\alias{validate_ready4fun_functions}
\title{validate ready4fun package functions Function types lookup table.}
\usage{
validate_ready4fun_functions(x)
}
\arguments{
\item{x}{An unvalidated instance of the Function types lookup table.}
}
\value{
A prototpe for Function types lookup table.
}
\description{
Validate an instance of the Function types lookup table.
}
\details{
Function types lookup table.
}
\keyword{internal}
