% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_metadata_a.R
\name{validate_ready4fun_metadata_a}
\alias{validate_ready4fun_metadata_a}
\title{validate ready4fun package metadata a ready4 submodule class for package metadata required for initial package set-up step.}
\usage{
validate_ready4fun_metadata_a(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 submodule class for package metadata required for initial package set-up step.}
}
\value{
A prototpe for ready4 submodule class for package metadata required for initial package set-up step.
}
\description{
Validate an instance of the ready4 submodule class for package metadata required for initial package set-up step.
}
\details{
Package metadata (A).
}
\keyword{internal}
