% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_objects.R
\name{validate_ready4fun_objects}
\alias{validate_ready4fun_objects}
\title{validate ready4fun package objects Object types lookup table.}
\usage{
validate_ready4fun_objects(x)
}
\arguments{
\item{x}{An unvalidated instance of the Object types lookup table.}
}
\value{
A prototpe for Object types lookup table.
}
\description{
Validate an instance of the Object types lookup table.
}
\details{
Object types lookup table.
}
\keyword{internal}
