% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_desc.R
\name{validate_ready4fun_pkg_desc}
\alias{validate_ready4fun_pkg_desc}
\title{Validate ready4fun package package description ready4 S4 class for declaring package description file data.}
\usage{
validate_ready4fun_pkg_desc(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S4 class for declaring package description file data.}
}
\value{
A prototpe for ready4 S4 class for declaring package description file data.
}
\description{
Validate an instance of the ready4 S4 class for declaring package description file data.
}
\details{
ready4 S3 class for declaring package description file data.
}
