% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_ds.R
\name{validate_ready4fun_pkg_ds}
\alias{validate_ready4fun_pkg_ds}
\title{Validate ready4fun package dataset ready4 S4 class for declaring package description file data..1}
\usage{
validate_ready4fun_pkg_ds(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S4 class for declaring package description file data..1}
}
\value{
A prototpe for ready4 S4 class for declaring package description file data..1
}
\description{
Validate an instance of the ready4 S4 class for declaring package description file data..1
}
\details{
ready4 S3 class for declaring package description file data.
}
