% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_setup.R
\name{validate_ready4fun_pkg_setup}
\alias{validate_ready4fun_pkg_setup}
\title{Validate ready4 S4 class for package metadata required for package set-up.}
\usage{
validate_ready4fun_pkg_setup(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S4 class for package metadata required for package set-up.}
}
\value{
A prototpe for ready4 S4 class for package metadata required for package set-up.
}
\description{
Validate an instance of the ready4 S4 class for package metadata required for package set-up.
}
\details{
ready4 S3 class for package metadata required for package set-up.
}
