% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_pkg_setup_two.R
\name{validate_ready4fun_pkg_setup_two}
\alias{validate_ready4fun_pkg_setup_two}
\title{Validate ready4fun package setup two ready4 S4 class for package metadata required for second package set-up step.}
\usage{
validate_ready4fun_pkg_setup_two(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S4 class for package metadata required for second package set-up step.}
}
\value{
A prototpe for ready4 S4 class for package metadata required for second package set-up step.
}
\description{
Validate an instance of the ready4 S4 class for package metadata required for second package set-up step.
}
\details{
ready4 S3 class for package metadata required for second package set-up step.
}
