% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_abbr_lup}
\alias{write_abbr_lup}
\title{Write abbreviation}
\usage{
write_abbr_lup(
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  no_plural_chr = NA_character_,
  custom_plural_ls = NULL,
  overwrite_1L_lgl = T,
  seed_lup = NULL,
  url_1L_chr,
  pkg_nm_1L_chr = get_dev_pkg_nm()
)
}
\arguments{
\item{short_name_chr}{Short name (a character vector), Default: 'NA'}

\item{long_name_chr}{Long name (a character vector), Default: 'NA'}

\item{no_plural_chr}{No plural (a character vector), Default: 'NA'}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: T}

\item{seed_lup}{Seed (a lookup table), Default: NULL}

\item{url_1L_chr}{Url (a character vector of length one)}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: get_dev_pkg_nm()}
}
\description{
write_abbr_lup() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write abbreviation lookup table.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
