% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_all_tbs_in_tbs_r4_to_csvs}
\alias{write_all_tbs_in_tbs_r4_to_csvs}
\title{Write all tibbles in tibbles readyforwhatsnext S4 to csvs}
\usage{
write_all_tbs_in_tbs_r4_to_csvs(tbs_r4, r4_name_chr, lup_dir_chr, pfx_chr)
}
\arguments{
\item{tbs_r4}{Tibbles (a readyforwhatsnext S4)}

\item{r4_name_chr}{Readyforwhatsnext S4 name (a character vector of length 1)}

\item{lup_dir_chr}{Lookup table directory (a character vector of length 1)}

\item{pfx_chr}{Prefix (a character vector of length 1)}
}
\description{
write_all_tbs_in_tbs_r4_to_csvs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write all a tibbles in a tibbles readyforwhatsnext S4 to csvs.NA
}
