% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_and_doc_ds}
\alias{write_and_doc_ds}
\title{Write and document dataset}
\usage{
write_and_doc_ds(
  db_df,
  overwrite_1L_lgl = T,
  db_1L_chr,
  title_1L_chr,
  desc_1L_chr,
  format_1L_chr = "A tibble",
  url_1L_chr = NA_character_,
  vars_ls = NULL,
  R_dir_1L_chr = "R",
  simple_lup_1L_lgl = F,
  abbreviations_lup = NULL,
  object_type_lup = NULL,
  pkg_dss_tb = tibble::tibble(ds_obj_nm_chr = character(0), title_chr = character(0),
    desc_chr = character(0), url_chr = character(0)),
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  server_1L_chr = deprecated()
)
}
\arguments{
\item{db_df}{Database (a data.frame)}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: T}

\item{db_1L_chr}{Database (a character vector of length one)}

\item{title_1L_chr}{Title (a character vector of length one)}

\item{desc_1L_chr}{Description (a character vector of length one)}

\item{format_1L_chr}{Format (a character vector of length one), Default: 'A tibble'}

\item{url_1L_chr}{Url (a character vector of length one), Default: 'NA'}

\item{vars_ls}{Variables (a list), Default: NULL}

\item{R_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}

\item{simple_lup_1L_lgl}{Simple lookup table (a logical vector of length one), Default: F}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{pkg_dss_tb}{Package datasets (a tibble), Default: tibble::tibble(ds_obj_nm_chr = character(0), title_chr = character(0),
desc_chr = character(0), url_chr = character(0))}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Package datasets (a tibble)
}
\description{
write_and_doc_ds() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write and document dataset. The function returns Package datasets (a tibble).
}
\keyword{internal}
