% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_and_doc_fn_fls}
\alias{write_and_doc_fn_fls}
\title{Write and document function files}
\usage{
write_and_doc_fn_fls(
  fns_dmt_tb,
  r_dir_1L_chr = "R",
  path_to_user_dmt_dir_1L_chr = "../../../../Documentation/Code/User",
  path_to_dvpr_dmt_dir_1L_chr = "../../../../Documentation/Code/Developer",
  make_pdfs_1L_lgl = T,
  dev_pkgs_chr = NA_character_
)
}
\arguments{
\item{fns_dmt_tb}{Functions documentation (a tibble)}

\item{r_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}

\item{path_to_user_dmt_dir_1L_chr}{Path to user documentation directory (a character vector of length one), Default: '../../../../Documentation/Code/User'}

\item{path_to_dvpr_dmt_dir_1L_chr}{Path to developer documentation directory (a character vector of length one), Default: '../../../../Documentation/Code/Developer'}

\item{make_pdfs_1L_lgl}{Make pdfs (a logical vector of length one), Default: T}

\item{dev_pkgs_chr}{Development packages (a character vector), Default: 'NA'}
}
\description{
write_and_doc_fn_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write and document function files.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
