% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_citation_fl}
\alias{write_citation_fl}
\title{Write citation file}
\usage{
write_citation_fl(pkg_setup_ls)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}
}
\description{
write_citation_fl() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write citation file. The function is called for its side effects and does not return a value.
}
\keyword{internal}
