% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_dmtd_fn_type_lup}
\alias{write_dmtd_fn_type_lup}
\title{Write documented function type lookup table}
\usage{
write_dmtd_fn_type_lup(
  fn_types_lup = make_fn_type_lup(),
  overwrite_1L_lgl = T,
  pkg_nm_1L_chr = get_dev_pkg_nm(),
  url_1L_chr = deprecated(),
  abbreviations_lup = NULL,
  object_type_lup = NULL,
  pkg_dss_tb = tibble::tibble(ds_obj_nm_chr = character(0), title_chr = character(0),
    desc_chr = character(0), url_chr = character(0)),
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  server_1L_chr = deprecated()
)
}
\arguments{
\item{fn_types_lup}{Function types (a lookup table), Default: make_fn_type_lup()}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: T}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: get_dev_pkg_nm()}

\item{url_1L_chr}{Url (a character vector of length one), Default: deprecated()}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{pkg_dss_tb}{Package datasets (a tibble), Default: tibble::tibble(ds_obj_nm_chr = character(0), title_chr = character(0),
desc_chr = character(0), url_chr = character(0))}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
No return value, called for side effects.
}
\description{
write_dmtd_fn_type_lup() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write documented function type lookup table. The function is called for its side effects and does not return a value.
}
\keyword{internal}
