% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_documented_fns}
\alias{write_documented_fns}
\title{Write documented functions}
\usage{
write_documented_fns(tmp_fn_dir_1L_chr, R_dir_1L_chr)
}
\arguments{
\item{tmp_fn_dir_1L_chr}{Temporary function directory (a character vector of length one)}

\item{R_dir_1L_chr}{R directory (a character vector of length one)}
}
\value{
No return value, called for side effects.
}
\description{
write_documented_fns() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write documented functions. The function is called for its side effects and does not return a value.
}
\keyword{internal}
