% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_documented_fns_R}
\alias{write_documented_fns_R}
\title{Write documented functions}
\usage{
write_documented_fns_R(tmp_fn_dir_chr, R_dir_chr)
}
\arguments{
\item{tmp_fn_dir_chr}{Tmp function directory (a character vector of length 1)}

\item{R_dir_chr}{R directory (a character vector of length 1)}
}
\description{
write_documented_fns_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a documented functions R. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
