% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ds_dmt}
\alias{write_ds_dmt}
\title{Write dataset documentation}
\usage{
write_ds_dmt(
  db,
  db_1L_chr,
  title_1L_chr,
  desc_1L_chr,
  format_1L_chr = "A tibble",
  url_1L_chr = NA_character_,
  vars_ls = NULL,
  R_dir_1L_chr = "R",
  abbreviations_lup = NULL,
  object_type_lup = NULL
)
}
\arguments{
\item{db}{PARAM_DESCRIPTION}

\item{db_1L_chr}{Database (a character vector of length one)}

\item{title_1L_chr}{Title (a character vector of length one)}

\item{desc_1L_chr}{Description (a character vector of length one)}

\item{format_1L_chr}{Format (a character vector of length one), Default: 'A tibble'}

\item{url_1L_chr}{Url (a character vector of length one), Default: 'NA'}

\item{vars_ls}{Vars (a list), Default: NULL}

\item{R_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}
}
\description{
write_ds_dmt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write dataset documentation. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
