% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ds_dmt_R}
\alias{write_ds_dmt_R}
\title{Write dataset documentation}
\usage{
write_ds_dmt_R(
  db,
  db_chr,
  title_chr,
  desc_chr,
  format_chr = "A tibble",
  url_chr = NA_character_,
  vars_ls = NULL,
  R_dir_chr = "R",
  abbreviations_lup = NULL,
  object_type_lup = NULL
)
}
\arguments{
\item{db}{PARAM_DESCRIPTION}

\item{db_chr}{Database (a character vector of length 1)}

\item{title_chr}{Title (a character vector of length 1)}

\item{desc_chr}{Description (a character vector of length 1)}

\item{format_chr}{Format (a character vector of length 1), Default: 'A tibble'}

\item{url_chr}{Url (a character vector of length 1), Default: 'NA'}

\item{vars_ls}{Vars (a list), Default: NULL}

\item{R_dir_chr}{R directory (a character vector of length 1), Default: 'R'}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}
}
\description{
write_ds_dmt_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a dataset documentation R. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
