% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fn_dmt}
\alias{write_fn_dmt}
\title{Write function dmt}
\usage{
write_fn_dmt(
  fn_name_chr,
  fn_type_chr,
  fn = NULL,
  fn_desc_chr = NA_character_,
  fn_out_type_chr = NA_character_,
  fn_title_chr = NA_character_,
  example_lgl = F,
  export_lgl = T,
  class_name_chr = "",
  details_chr = "DETAILS",
  args_ls = NULL,
  import_chr_vec = NA_character_,
  doc_in_class_lgl = F,
  object_type_lup = NULL
)
}
\arguments{
\item{fn_name_chr}{Function name (a character vector of length 1)}

\item{fn_type_chr}{Function type (a character vector of length 1)}

\item{fn}{Function (a function), Default: NULL}

\item{fn_desc_chr}{Function description (a character vector of length 1), Default: 'NA'}

\item{fn_out_type_chr}{Function out type (a character vector of length 1), Default: 'NA'}

\item{fn_title_chr}{Function title (a character vector of length 1), Default: 'NA'}

\item{example_lgl}{Example (a logical vector of length 1), Default: F}

\item{export_lgl}{Export (a logical vector of length 1), Default: T}

\item{class_name_chr}{Class name (a character vector of length 1), Default: ''}

\item{details_chr}{Details (a character vector of length 1), Default: 'DETAILS'}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{import_chr_vec}{Import (a character vector), Default: 'NA'}

\item{doc_in_class_lgl}{Document in class (a logical vector of length 1), Default: F}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}
}
\description{
write_fn_dmt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a function dmt. The function is called for its side effects and does not return a value.
}
