% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fn_fl_R}
\alias{write_fn_fl_R}
\title{Write function file}
\usage{
write_fn_fl_R(fns_dmt_tb, r_dir_chr = "R", document_unexp_lgl = T)
}
\arguments{
\item{fns_dmt_tb}{Functions dmt (a tibble)}

\item{r_dir_chr}{R directory (a character vector of length 1), Default: 'R'}

\item{document_unexp_lgl}{Document unexported (a logical vector of length 1), Default: T}
}
\description{
write_fn_fl_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a function file R.NA
}
\keyword{internal}
