% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fn_type_dirs}
\alias{write_fn_type_dirs}
\title{Write function type directories}
\usage{
write_fn_type_dirs(path_1L_chr = "data-raw")
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one), Default: 'data-raw'}
}
\description{
write_fn_type_dirs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a function type directories.NA
}
\keyword{internal}
