% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fns_dmt_tb}
\alias{write_fns_dmt_tb}
\title{Write functions documentation tibble}
\usage{
write_fns_dmt_tb(
  pkg_setup_ls,
  gh_prerelease_1L_lgl = T,
  gh_repo_desc_1L_chr = "Supplementary Files",
  gh_tag_1L_chr = "Documentation_0.0"
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{gh_prerelease_1L_lgl}{Github prerelease (a logical vector of length one), Default: T}

\item{gh_repo_desc_1L_chr}{Github repository description (a character vector of length one), Default: 'Supplementary Files'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}
}
\value{
No return value, called for side effects.
}
\description{
write_fns_dmt_tb() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write functions documentation tibble. The function is called for its side effects and does not return a value.
}
\keyword{internal}
