% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_from_tmp}
\alias{write_from_tmp}
\title{Write from temporary}
\usage{
write_from_tmp(
  temp_path_1L_chr,
  dest_path_1L_chr,
  edit_fn = function(x) {     x },
  args_ls = NULL
)
}
\arguments{
\item{temp_path_1L_chr}{Temp path (a character vector of length one)}

\item{dest_path_1L_chr}{Dest path (a character vector of length one)}

\item{edit_fn}{Edit (a function), Default: function(x) {
x
}}

\item{args_ls}{Arguments (a list), Default: NULL}
}
\description{
write_from_tmp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write from temporary.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
