% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_from_tmp_R}
\alias{write_from_tmp_R}
\title{Write from tmp}
\usage{
write_from_tmp_R(
  temp_path_chr,
  dest_path_chr,
  edit_fn = function(x) {     x },
  args_ls = NULL
)
}
\arguments{
\item{temp_path_chr}{Temp path (a character vector of length 1)}

\item{dest_path_chr}{Dest path (a character vector of length 1)}

\item{edit_fn}{Edit (a function), Default: function(x) {
   x
}}

\item{args_ls}{Arguments (a list), Default: NULL}
}
\description{
write_from_tmp_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write from tmp R. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
