% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_inst_dir}
\alias{write_inst_dir}
\title{Write instance directory}
\usage{
write_inst_dir(path_to_pkg_rt_1L_chr = getwd())
}
\arguments{
\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: getwd()}
}
\description{
write_inst_dir() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write instance directory. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
