% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_links_for_website}
\alias{write_links_for_website}
\title{Write links for website}
\usage{
write_links_for_website(
  path_to_pkg_rt_1L_chr = getwd(),
  user_manual_url_1L_chr = NA_character_,
  developer_manual_url_1L_chr = NA_character_,
  project_website_url_1L_chr = NA_character_
)
}
\arguments{
\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: getwd()}

\item{user_manual_url_1L_chr}{User manual url (a character vector of length one), Default: 'NA'}

\item{developer_manual_url_1L_chr}{Developer manual url (a character vector of length one), Default: 'NA'}

\item{project_website_url_1L_chr}{Project website url (a character vector of length one), Default: 'NA'}
}
\description{
write_links_for_website() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write links for website. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
