% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_new_abbrs}
\alias{write_new_abbrs}
\title{Write new abbreviations}
\usage{
write_new_abbrs(
  pkg_setup_ls,
  long_name_chr = NULL,
  custom_plural_ls = NULL,
  key_1L_chr = deprecated(),
  no_plural_chr = NA_character_,
  publish_dv_1L_lgl = deprecated(),
  pfx_rgx = NA_character_,
  server_1L_chr = deprecated()
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{long_name_chr}{Long name (a character vector), Default: NULL}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{no_plural_chr}{No plural (a character vector), Default: 'NA'}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: deprecated()}

\item{pfx_rgx}{Prefix (a regular expression vector), Default: 'NA'}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Package setup (a list)
}
\description{
write_new_abbrs() is a Write function that writes a file to a specified local directory Specifically, this function implements an algorithm to write new abbreviations. The function returns Package setup (a list).
}
\keyword{internal}
