% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_new_arg_sfxs}
\alias{write_new_arg_sfxs}
\title{Write new argument sfxs}
\usage{
write_new_arg_sfxs(
  arg_nms_chr,
  fn_type_1L_chr,
  dir_path_chr,
  rt_dev_dir_path_1L_chr = normalizePath("../../../"),
  pkg_nm_1L_chr,
  inc_fns_idx_dbl = NA_real_
)
}
\arguments{
\item{arg_nms_chr}{Argument names (a character vector)}

\item{fn_type_1L_chr}{Function type (a character vector of length one)}

\item{dir_path_chr}{Directory path (a character vector)}

\item{rt_dev_dir_path_1L_chr}{Root development directory path (a character vector of length one), Default: normalizePath("../../../")}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one)}

\item{inc_fns_idx_dbl}{Include functions index (a double vector), Default: NA}
}
\value{
Function arguments to rnm (a list)
}
\description{
write_new_arg_sfxs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write new argument sfxs. The function returns Function arguments to rnm (a list).
}
