% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_new_fn_types}
\alias{write_new_fn_types}
\title{Write new function types}
\usage{
write_new_fn_types(
  pkg_setup_ls,
  fn_type_desc_chr = NA_character_,
  first_arg_desc_chr = NA_character_,
  is_generic_lgl = F,
  is_method_lgl = F,
  key_1L_chr = deprecated(),
  second_arg_desc_chr = NA_character_,
  server_1L_chr = deprecated(),
  publish_dv_1L_lgl = deprecated()
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{fn_type_desc_chr}{Function type description (a character vector), Default: 'NA'}

\item{first_arg_desc_chr}{First argument description (a character vector), Default: 'NA'}

\item{is_generic_lgl}{Is generic (a logical vector), Default: F}

\item{is_method_lgl}{Is method (a logical vector), Default: F}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{second_arg_desc_chr}{Second argument description (a character vector), Default: 'NA'}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: deprecated()}
}
\value{
Package setup (a list)
}
\description{
write_new_fn_types() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write new function types. The function returns Package setup (a list).
}
