% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_new_obj_types}
\alias{write_new_obj_types}
\title{Write new object types}
\usage{
write_new_obj_types(
  pkg_setup_ls,
  long_name_chr = NULL,
  atomic_element_lgl = F,
  custom_plural_ls = NULL,
  key_1L_chr = deprecated(),
  no_plural_chr = NA_character_,
  publish_dv_1L_lgl = deprecated(),
  pfx_rgx = NA_character_,
  r3_can_extend_lgl = F,
  server_1L_chr = deprecated()
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{long_name_chr}{Long name (a character vector), Default: NULL}

\item{atomic_element_lgl}{Atomic element (a logical vector), Default: F}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{no_plural_chr}{No plural (a character vector), Default: 'NA'}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: deprecated()}

\item{pfx_rgx}{Prefix (a regular expression vector), Default: 'NA'}

\item{r3_can_extend_lgl}{Ready4 S3 can extend (a logical vector), Default: F}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Package setup (a list)
}
\description{
write_new_obj_types() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write new object types. The function returns Package setup (a list).
}
