% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_new_words_vec}
\alias{write_new_words_vec}
\title{Write new words vector}
\usage{
write_new_words_vec(
  pkg_setup_ls,
  key_1L_chr = deprecated(),
  publish_dv_1L_lgl = deprecated()
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: deprecated()}
}
\value{
Package setup (a list)
}
\description{
write_new_words_vec() is a Write function that writes a file to a specified local directory Specifically, this function implements an algorithm to write new words vector. The function returns Package setup (a list).
}
\keyword{internal}
