% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ns_imps_to_desc}
\alias{write_ns_imps_to_desc}
\title{Write namespace imports to description}
\usage{
write_ns_imps_to_desc(dev_pkgs_chr = NA_character_, incr_ver_1L_lgl = T)
}
\arguments{
\item{dev_pkgs_chr}{Development packages (a character vector), Default: 'NA'}

\item{incr_ver_1L_lgl}{Incr ver (a logical vector of length one), Default: T}
}
\description{
write_ns_imps_to_desc() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a namespace imports to a description.NA
}
\keyword{internal}
