% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_package}
\alias{write_package}
\title{Write package}
\usage{
write_package(
  pkg_setup_ls,
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  publish_dv_1L_lgl = T,
  self_serve_1L_lgl = F,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  cls_fn_ls = deprecated(),
  path_to_dmt_dir_1L_chr = deprecated(),
  pkg_desc_ls = deprecated(),
  pkg_ds_ls_ls = deprecated()
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: T}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{cls_fn_ls}{Class (a list of functions), Default: deprecated()}

\item{path_to_dmt_dir_1L_chr}{Path to documentation directory (a character vector of length one), Default: deprecated()}

\item{pkg_desc_ls}{Package description (a list), Default: deprecated()}

\item{pkg_ds_ls_ls}{Package dataset (a list of lists), Default: deprecated()}
}
\description{
write_package() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write package. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
