% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_pkg}
\alias{write_pkg}
\title{Write package}
\usage{
write_pkg(package_1L_chr, R_dir_1L_chr = "R")
}
\arguments{
\item{package_1L_chr}{Package (a character vector of length one)}

\item{R_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}
}
\description{
write_pkg() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to Write package.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
