% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_pkg_R}
\alias{write_pkg_R}
\title{Write package}
\usage{
write_pkg_R(package_chr, R_dir_chr = "R")
}
\arguments{
\item{package_chr}{Package (a character vector of length 1)}

\item{R_dir_chr}{R directory (a character vector of length 1), Default: 'R'}
}
\description{
write_pkg_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a package R.NA
}
