% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_pkg_setup_fls}
\alias{write_pkg_setup_fls}
\title{Write package setup files}
\usage{
write_pkg_setup_fls(
  path_to_pkg_rt_1L_chr = ".",
  dev_pkg_nm_1L_chr = get_dev_pkg_nm(),
  make_tmpl_vignette_1L_lgl = F,
  incr_ver_1L_lgl = T
)
}
\arguments{
\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: '.'}

\item{dev_pkg_nm_1L_chr}{Development package name (a character vector of length one), Default: get_dev_pkg_nm()}

\item{make_tmpl_vignette_1L_lgl}{Make tmpl vignette (a logical vector of length one), Default: F}

\item{incr_ver_1L_lgl}{Incr ver (a logical vector of length one), Default: T}
}
\description{
write_pkg_setup_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a package setup files.NA
}
