% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_pkg_setup_fls_R}
\alias{write_pkg_setup_fls_R}
\title{Write package setup files}
\usage{
write_pkg_setup_fls_R(
  path_to_pkg_rt_chr = ".",
  dev_pkg_nm_chr = NA_character_,
  make_tmpl_vignette_lgl = F,
  incr_ver_lgl = T
)
}
\arguments{
\item{path_to_pkg_rt_chr}{Path to package rt (a character vector of length 1), Default: '.'}

\item{dev_pkg_nm_chr}{Dev package name (a character vector of length 1), Default: 'NA'}

\item{make_tmpl_vignette_lgl}{Make tmpl vignette (a logical vector of length 1), Default: F}

\item{incr_ver_lgl}{Incr ver (a logical vector of length 1), Default: T}
}
\description{
write_pkg_setup_fls_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a package setup files R.NA
}
\keyword{internal}
