% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_pt_lup_db_R}
\alias{write_pt_lup_db_R}
\title{Write prototype lookup table database}
\usage{
write_pt_lup_db_R(R_dir_chr = "R")
}
\arguments{
\item{R_dir_chr}{R directory (a character vector of length 1), Default: 'R'}
}
\description{
write_pt_lup_db_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a prototype lookup table database R.NA
}
