% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_std_imp}
\alias{write_std_imp}
\title{Write standard import}
\usage{
write_std_imp(R_dir_1L_chr = "R")
}
\arguments{
\item{R_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}
}
\description{
write_std_imp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write standard import. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
