% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_std_imp_R}
\alias{write_std_imp_R}
\title{Write standard import}
\usage{
write_std_imp_R(R_dir_chr = "R")
}
\arguments{
\item{R_dir_chr}{R directory (a character vector of length 1), Default: 'R'}
}
\description{
write_std_imp_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a standard import R.NA
}
