% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_remove_collate}
\alias{write_to_remove_collate}
\title{Write to remove collate}
\usage{
write_to_remove_collate(description_chr)
}
\arguments{
\item{description_chr}{Description (a character vector)}
}
\value{
Description (a character vector)
}
\description{
write_to_remove_collate() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to remove collate.The function returns a description character vector.
}
