% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_replace_fn_nms}
\alias{write_to_replace_fn_nms}
\title{Write to replace function names}
\usage{
write_to_replace_fn_nms(
  rename_tb,
  undocumented_fns_dir_chr = make_undmtd_fns_dir_chr(),
  rt_dev_dir_path_1L_chr = normalizePath("../../../"),
  dev_pkg_nm_1L_chr = get_dev_pkg_nm()
)
}
\arguments{
\item{rename_tb}{Rename (a tibble)}

\item{undocumented_fns_dir_chr}{Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr()}

\item{rt_dev_dir_path_1L_chr}{Root development directory path (a character vector of length one), Default: normalizePath("../../../")}

\item{dev_pkg_nm_1L_chr}{Development package name (a character vector of length one), Default: get_dev_pkg_nm()}
}
\description{
write_to_replace_fn_nms() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to replace function names.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
