% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_rpl_1L_and_indefL_sfcs}
\alias{write_to_rpl_1L_and_indefL_sfcs}
\title{Write to replace length one and indefinite length suffices}
\usage{
write_to_rpl_1L_and_indefL_sfcs(
  indefL_arg_nm_1L_chr,
  file_path_1L_chr = NA_character_,
  dir_path_1L_chr = NA_character_
)
}
\arguments{
\item{indefL_arg_nm_1L_chr}{Indefinite length argument name (a character vector of length one)}

\item{file_path_1L_chr}{File path (a character vector of length one), Default: 'NA'}

\item{dir_path_1L_chr}{Directory path (a character vector of length one), Default: 'NA'}
}
\description{
write_to_rpl_1L_and_indefL_sfcs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to Write to replace length one and indefinite length suffices.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
