% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_tidy_pkg}
\alias{write_to_tidy_pkg}
\title{Write to tidy package}
\usage{
write_to_tidy_pkg(
  pkg_setup_ls,
  build_vignettes_1L_lgl = TRUE,
  clean_license_1L_lgl = TRUE,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  examples_chr = character(0),
  options_chr = c("Y", "N"),
  project_1L_chr = "Model",
  suggest_chr = "pkgload"
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{build_vignettes_1L_lgl}{Build vignettes (a logical vector of length one), Default: TRUE}

\item{clean_license_1L_lgl}{Clean license (a logical vector of length one), Default: TRUE}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{examples_chr}{Examples (a character vector), Default: character(0)}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{project_1L_chr}{Project (a character vector of length one), Default: 'Model'}

\item{suggest_chr}{Suggest (a character vector), Default: 'pkgload'}
}
\value{
No return value, called for side effects.
}
\description{
write_to_tidy_pkg() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to tidy package. The function is called for its side effects and does not return a value.
}
\keyword{internal}
