% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_vignette_R}
\alias{write_vignette_R}
\title{Write vignette}
\usage{
write_vignette_R(package_chr, pkg_rt_dir_chr = ".")
}
\arguments{
\item{package_chr}{Package (a character vector of length 1)}

\item{pkg_rt_dir_chr}{Package rt directory (a character vector of length 1), Default: '.'}
}
\description{
write_vignette_R() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write vignette R.NA
}
\keyword{internal}
