% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_addl_pkgs}
\alias{add_addl_pkgs}
\title{Add additional packages}
\usage{
add_addl_pkgs(addl_pkgs_ls)
}
\arguments{
\item{addl_pkgs_ls}{Additional packages (a list)}
}
\description{
add_addl_pkgs() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add additional packages. Function argument addl_pkgs_ls specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
