% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_build_ignore}
\alias{add_build_ignore}
\title{Add build ignore}
\usage{
add_build_ignore(build_ignore_ls)
}
\arguments{
\item{build_ignore_ls}{Build ignore (a list)}
}
\description{
add_build_ignore() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add build ignore. The function is called for its side effects and does not return a value.
}
\keyword{internal}
