% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_indef_artl_to_item}
\alias{add_indef_artl_to_item}
\title{Add indefinite article to item}
\usage{
add_indef_artl_to_item(
  phrase_chr,
  abbreviations_lup = NULL,
  ignore_phrs_not_in_lup_1L_lgl = T
)
}
\arguments{
\item{phrase_chr}{Phrase (a character vector)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{ignore_phrs_not_in_lup_1L_lgl}{Ignore phrases not in lookup table (a logical vector of length one), Default: T}
}
\value{
Indefinite item (a character vector)
}
\description{
add_indef_artl_to_item() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to Add indefinite article to item. Function argument phrase_chr specifies the object to be updated.The function returns an indefinite item (a character vector).
}
