% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{analyse}
\alias{analyse}
\title{Analyse}
\usage{
analyse(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
analyse() is an Analyse function that runs an analysis or analyses using specified data and parameters. Specifically, this function implements an algorithm to analyse. The function is called for its side effects and does not return a value.
}
