% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_author.R
\name{author}
\alias{author}
\alias{author.ready4fun_pkg_setup}
\alias{author,ready4fun_pkg_setup-method}
\title{Author}
\usage{
author(x, ...)

author.ready4fun_pkg_setup(
  x,
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  list_generics_1L_lgl = F,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)

\S4method{author}{ready4fun_pkg_setup}(
  x,
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  list_generics_1L_lgl = F,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{x}{An instance of ready4 S4 class for package metadata required for package set-up.}

\item{...}{Additional arguments (an additional arguments)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{list_generics_1L_lgl}{List generics (a logical vector of length one), Default: F}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{self_serve_fn_ls}{Self serve (a list of functions), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
Ready4fun package package setup (a ready4 S3)
}
\description{
author() is an Author function that writes files to local or remote locations. Specifically, this function implements an algorithm to author.

author.ready4fun_pkg_setup() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S4 class for package metadata required for package set-up.. The function returns Ready4fun package package setup (a ready4 S3).
}
