% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_author.R
\name{author}
\alias{author}
\alias{author.ready4fun_manifest}
\alias{author,ready4fun_manifest-method}
\alias{author.ready4fun_metadata_a}
\alias{author,ready4fun_metadata_a-method}
\title{Author}
\usage{
author(x, ...)

author.ready4fun_manifest(
  x,
<<<<<<< HEAD
=======
  dv_url_pfx_1L_chr = character(0),
>>>>>>> dev
  key_1L_chr = NULL,
  list_generics_1L_lgl = F,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL
)

\S4method{author}{ready4fun_manifest}(
  x,
<<<<<<< HEAD
=======
  dv_url_pfx_1L_chr = character(0),
>>>>>>> dev
  key_1L_chr = NULL,
  list_generics_1L_lgl = F,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL
)

author.ready4fun_metadata_a(x)

\S4method{author}{ready4fun_metadata_a}(x)
}
\arguments{
\item{x}{An instance of ready4 S3 class for package metadata required for initial package set-up step.}

\item{...}{Additional arguments (an additional arguments)}

<<<<<<< HEAD
=======
\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: character(0)}

>>>>>>> dev
\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{list_generics_1L_lgl}{List generics (a logical vector of length one), Default: F}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{self_serve_fn_ls}{Self serve (a list of functions), Default: NULL}
}
\value{
NA ()
}
\description{
author() is an Author function that writes files to local or remote locations. Specifically, this function implements an algorithm to author. The function is called for its side effects and does not return a value.

author.ready4fun_manifest() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.

author.ready4fun_metadata_a() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S3 class for package metadata required for initial package set-up step. The function is called for its side effects and does not return a value.
}
