% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author.ready4fun_manifest}
\alias{author.ready4fun_manifest}
\alias{author,ready4fun_manifest-method}
\alias{author.ready4fun_metadata_a}
\alias{author,ready4fun_metadata_a-method}
\title{Author method applied to ready4 S3 class for encapsulating the metadata required for package set-up..}
\usage{
author.ready4fun_manifest(
  x,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  list_generics_1L_lgl = T,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL
)

\S4method{author}{ready4fun_manifest}(
  x,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  list_generics_1L_lgl = T,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL
)

author.ready4fun_metadata_a(x)

\S4method{author}{ready4fun_metadata_a}(x)
}
\arguments{
\item{x}{An instance of ready4 S3 class for package metadata required for initial package set-up step.}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{list_generics_1L_lgl}{List generics (a logical vector of length one), Default: T}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{self_serve_fn_ls}{Self serve (a list of functions), Default: NULL}
}
\value{
x (An object)
}
\description{
author.ready4fun_manifest() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.

author.ready4fun_metadata_a() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S3 class for package metadata required for initial package set-up step. The function is called for its side effects and does not return a value.
}
