% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorClasses.R
\name{authorClasses.ready4fun_manifest}
\alias{authorClasses.ready4fun_manifest}
\alias{authorClasses,ready4fun_manifest-method}
\title{Author and document classes}
\usage{
authorClasses.ready4fun_manifest(
  x,
  key_1L_chr = NULL,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL
)

\S4method{authorClasses}{ready4fun_manifest}(
  x,
  key_1L_chr = NULL,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL
)
}
\arguments{
\item{x}{An instance of \code{ready4fun_manifest}, a ready4 submodule class for encapsulating the metadata required for package set-up.}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{self_serve_fn_ls}{Self serve (a list of functions), Default: NULL}
}
\value{
No return value, called for side effects.
}
\description{
authorClasses.ready4fun_manifest() is an authorClasses method that authors and saves files necessary for creating and documenting classes. This method is implemented for the ready4 submodule class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
