% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorData.R
\name{authorData.ready4fun_manifest}
\alias{authorData.ready4fun_manifest}
\alias{authorData,ready4fun_manifest-method}
\title{Author and document datasets}
\usage{
authorData.ready4fun_manifest(x)

\S4method{authorData}{ready4fun_manifest}(x)
}
\arguments{
\item{x}{An instance of \code{ready4fun_manifest}, a ready4 submodule class for encapsulating the metadata required for package set-up.}
}
\value{
x (An object)
}
\description{
authorData.ready4fun_manifest() is an authorData method that authors and saves files necessary for creating and documenting datasets. This method is implemented for the ready4 submodule class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
