% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorReport.R
\name{authorReport.ready4fun_manifest}
\alias{authorReport.ready4fun_manifest}
\alias{authorReport,ready4fun_manifest-method}
\title{Author and save a report}
\usage{
authorReport.ready4fun_manifest(x, key_1L_chr = Sys.getenv("DATAVERSE_KEY"))

\S4method{authorReport}{ready4fun_manifest}(x, key_1L_chr = Sys.getenv("DATAVERSE_KEY"))
}
\arguments{
\item{x}{An instance of \code{ready4fun_manifest}, a ready4 submodule class for encapsulating the metadata required for package set-up.}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}
}
\value{
No return value, called for side effects.
}
\description{
authorReport.ready4fun_manifest() is an authorReport method that authors and saves a report. This method is implemented for the ready4 submodule class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
