% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_abbrs}
\alias{get_abbrs}
\title{Get house style abbreviations}
\usage{
get_abbrs(
  what_1L_chr = character(0),
  type_1L_chr = c("abbreviation", "extension"),
  abbreviations_lup = NULL,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  dv_nm_1L_chr = NA_character_,
  dv_ds_metadata_ls = list(list()),
  dv_ds_nm_1L_chr = NA_character_,
  dv_server_1L_chr = NA_character_,
  dv_url_pfx_1L_chr = NA_character_,
  search_descs_1L_lgl = deprecated()
)
}
\arguments{
\item{what_1L_chr}{What (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("abbreviation", "extension")}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{dv_nm_1L_chr}{Dataverse name (a character vector of length one), Default: 'NA'}

\item{dv_ds_metadata_ls}{Dataverse dataset metadata (a list), Default: list(list())}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'NA'}

\item{dv_server_1L_chr}{Dataverse server (a character vector of length one), Default: 'NA'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: 'NA'}

\item{search_descs_1L_lgl}{Search descriptions (a logical vector of length one), Default: deprecated()}
}
\value{
Abbreviations (a lookup table)
}
\description{
An aspect of the ready4 framework is a consistent house style for code. Retrieve details on framework abbreviations with \code{get_abbrs}.
}
\examples{
library(ready4)
# Get current table of abbreviations
abbreviations_lup <- get_abbrs(gh_repo_1L_chr = "ready4-dev/ready4")
abbreviations_lup \%>\% head()
# Find abbreviations for a word
get_abbrs("template", abbreviations_lup = abbreviations_lup)
# Find possible whole word meanings associated with an abbreviation
get_abbrs("org", type_1L_chr = "extension", abbreviations_lup = abbreviations_lup)


}
