% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_abbrs}
\alias{get_abbrs}
\title{Get abbreviations}
\usage{
get_abbrs(text_1L_chr, abbreviations_lup = NULL, search_descs_1L_lgl = T)
}
\arguments{
\item{text_1L_chr}{Text (a character vector of length one)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{search_descs_1L_lgl}{Search descriptions (a logical vector of length one), Default: T}
}
\value{
Abbreviations (a lookup table)
}
\description{
get_abbrs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get abbreviations. Function argument text_1L_chr specifies the where to look for the required object. The function returns Abbreviations (a lookup table).
}
\keyword{internal}
