% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_all_depcys_of_fns}
\alias{get_all_depcys_of_fns}
\title{get all dependencies of functions}
\usage{
get_all_depcys_of_fns(pkg_depcy_ls, fns_chr)
}
\arguments{
\item{pkg_depcy_ls}{Package dependency (a list)}

\item{fns_chr}{Functions (a character vector)}
}
\value{
Functions to keep (a character vector)
}
\description{
get_all_depcys_of_fns() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get all dependencies of functions. Function argument pkg_depcy_ls specifies the where to look for the required object. The function returns Functions to keep (a character vector).
}
\keyword{internal}
