% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_arg_obj_type}
\alias{get_arg_obj_type}
\title{Get argument object type}
\usage{
get_arg_obj_type(
  argument_nm_1L_chr,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  object_type_lup = NULL,
  server_1L_chr = deprecated()
)
}
\arguments{
\item{argument_nm_1L_chr}{Argument name (a character vector of length one)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Argument object type (a character vector of length one)
}
\description{
get_arg_obj_type() is a Get function that extracts data from an object Specifically, this function implements an algorithm to get argument object type. The function returns Argument object type (a character vector of length one).
}
\keyword{internal}
