% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_arg_obj_type}
\alias{get_arg_obj_type}
\title{Get argument object type}
\usage{
get_arg_obj_type(argument_nm_1L_chr, object_type_lup = NULL)
}
\arguments{
\item{argument_nm_1L_chr}{Argument name (a character vector of length one)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}
}
\value{
Argument object type (a character vector of length one)
}
\description{
get_arg_obj_type() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get argument object type. Function argument argument_nm_1L_chr specifies the where to look for the required object. The function returns Argument object type (a character vector of length one).
}
\keyword{internal}
