% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dev_pkg_nm}
\alias{get_dev_pkg_nm}
\title{Get development package name}
\usage{
get_dev_pkg_nm(path_to_pkg_rt_1L_chr = ".")
}
\arguments{
\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: '.'}
}
\value{
Development package name (a character vector of length one)
}
\description{
get_dev_pkg_nm() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to Get development package name. Function argument path_to_pkg_rt_1L_chr specifies the where to look for the required object.The function returns a development package name (a character vector of length one).
}
