% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fn_args}
\alias{get_fn_args}
\title{Get function arguments}
\usage{
get_fn_args(fn)
}
\arguments{
\item{fn}{Function (a function)}
}
\value{
Function arguments (a character vector)
}
\description{
get_fn_args() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to Get function arguments. Function argument fn specifies the where to look for the required object.The function returns a function arguments (a character vector).
}
