% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fn_nms_in_file}
\alias{get_fn_nms_in_file}
\title{Get function names in file}
\usage{
get_fn_nms_in_file(path_1L_chr)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}
}
\value{
Local (a character vector)
}
\description{
get_fn_nms_in_file() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get function names in file. Function argument path_1L_chr specifies the where to look for the required object. The function returns Local (a character vector).
}
\keyword{internal}
