% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fn_types}
\alias{get_fn_types}
\title{Get supported function types}
\usage{
get_fn_types(
  dv_nm_1L_chr = NA_character_,
  dv_ds_metadata_ls = list(list()),
  dv_ds_nm_1L_chr = NA_character_,
  dv_server_1L_chr = NA_character_,
  dv_url_pfx_1L_chr = NA_character_,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  type_1L_chr = c("submodule", "simple", "partial"),
  what_1L_chr = c("regular", "method", "all")
)
}
\arguments{
\item{dv_nm_1L_chr}{Dataverse name (a character vector of length one), Default: 'NA'}

\item{dv_ds_metadata_ls}{Dataverse dataset metadata (a list), Default: list(list())}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'NA'}

\item{dv_server_1L_chr}{Dataverse server (a character vector of length one), Default: 'NA'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: 'NA'}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("submodule", "simple", "partial")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("regular", "method", "all")}
}
\value{
Function types (a lookup table)
}
\description{
All functions authored with the \code{ready4} framework need to begin with a verb. The definitions of all meaningful verbs used in functions authored for a ready4 framework model implementation can be retrieved using \code{get_fn_types()}.
}
