% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_cls_pts}
\alias{get_new_cls_pts}
\title{Get new class prototypes}
\usage{
get_new_cls_pts(pkg_setup_ls)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}
}
\value{
New class prototypes (a character vector)
}
\description{
get_new_cls_pts() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get new class prototypes. Function argument pkg_setup_ls specifies the where to look for the required object. The function returns New class prototypes (a character vector).
}
