% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_fn_types}
\alias{get_new_fn_types}
\title{Get new function types}
\usage{
get_new_fn_types(
  pkg_setup_ls,
  fn_nms_ls = make_fn_nms(),
  undmtd_fns_dir_chr = make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{fn_nms_ls}{Function names (a list), Default: make_fn_nms()}

\item{undmtd_fns_dir_chr}{Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)}
}
\value{
New function types (a character vector)
}
\description{
get_new_fn_types() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get new function types. Function argument pkg_setup_ls specifies the where to look for the required object. The function returns New function types (a character vector).
}
\keyword{internal}
