% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_fn_types}
\alias{get_new_fn_types}
\title{Get new function types}
\usage{
get_new_fn_types(
  abbreviations_lup,
  fn_type_lup_tb,
  fn_nms_ls = make_fn_nms(),
  undmtd_fns_dir_chr = make_undmtd_fns_dir_chr()
)
}
\arguments{
\item{abbreviations_lup}{Abbreviations (a lookup table)}

\item{fn_type_lup_tb}{Function type lookup table (a tibble)}

\item{fn_nms_ls}{Function names (a list), Default: make_fn_nms()}

\item{undmtd_fns_dir_chr}{Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr()}
}
\value{
New function types (a character vector)
}
\description{
get_new_fn_types() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get new function types. Function argument abbreviations_lup specifies the where to look for the required object. The function returns New function types (a character vector).
}
\keyword{internal}
