% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_outp_obj_type}
\alias{get_outp_obj_type}
\title{get output object type}
\usage{
get_outp_obj_type(
  fns_chr,
  abbreviations_lup,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  fns_env_ls,
  is_generic_lgl = F,
  key_1L_chr = deprecated(),
  object_type_lup = NULL,
  server_1L_chr = deprecated()
)
}
\arguments{
\item{fns_chr}{Functions (a character vector)}

\item{abbreviations_lup}{Abbreviations (a lookup table)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{fns_env_ls}{Functions (a list of environments)}

\item{is_generic_lgl}{Is generic (a logical vector), Default: F}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Output object type (a character vector)
}
\description{
get_outp_obj_type() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get output object type. Function argument fns_chr specifies the where to look for the required object. The function returns Output object type (a character vector).
}
\keyword{internal}
