% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_outp_obj_type}
\alias{get_outp_obj_type}
\title{Get output object type}
\usage{
get_outp_obj_type(fns_chr)
}
\arguments{
\item{fns_chr}{Functions (a character vector)}
}
\value{
Output object type (a character vector)
}
\description{
get_outp_obj_type() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get output object type. Function argument fns_chr specifies the where to look for the required object.The function returns an output object type (a character vector).
}
\keyword{internal}
