% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_return_obj_nm}
\alias{get_return_obj_nm}
\title{Get return object name}
\usage{
get_return_obj_nm(fn)
}
\arguments{
\item{fn}{Function (a function)}
}
\value{
Return (a character vector of length one)
}
\description{
get_return_obj_nm() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get return an object name. Function argument fn specifies the where to look for the required object.The function returns return (a character vector of length one).
}
