% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_abbreviations.R
\name{is_ready4fun_abbreviations}
\alias{is_ready4fun_abbreviations}
\title{Is ready4 S3 class for tibble object lookup table of abbreviations.}
\usage{
is_ready4fun_abbreviations(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table of abbreviations.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for tibble object lookup table of abbreviations.
}
\details{
ready4 S3 class for tibble object lookup table of abbreviations.
}
