% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_desc_spine}
\alias{make_arg_desc_spine}
\title{Make argument description spine}
\usage{
make_arg_desc_spine(
  argument_nm_1L_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL
)
}
\arguments{
\item{argument_nm_1L_chr}{Argument name (a character vector of length one)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\value{
NA ()
}
\description{
make_arg_desc_spine() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make argument description spine. The function is called for its side effects and does not return a value.
}
