% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_title}
\alias{make_arg_title}
\title{Make argument title}
\usage{
make_arg_title(
  args_chr,
  match_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL
)
}
\arguments{
\item{args_chr}{Arguments (a character vector)}

\item{match_chr}{Match (a character vector)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\value{
Title (a character vector)
}
\description{
make_arg_title() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make argument title.The function returns title character vector.
}
