% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_type}
\alias{make_arg_type}
\title{Make argument type}
\usage{
make_arg_type(
  fn_args_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL,
  fn
)
}
\arguments{
\item{fn_args_chr}{Function arguments (a character vector)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{fn}{Function (a function)}
}
\value{
Argument description (a character vector)
}
\description{
make_arg_type() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make argument type. The function returns an argument description (a character vector).
}
\keyword{internal}
