% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_type_abbr}
\alias{make_arg_type_abbr}
\title{Make argument type abbreviation}
\usage{
make_arg_type_abbr(
  fn_args_1L_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL
)
}
\arguments{
\item{fn_args_1L_chr}{Function arguments (a character vector of length one)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\description{
make_arg_type_abbr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make an argument type an abbreviation.NA
}
\keyword{internal}
