% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_type_abbr_spine}
\alias{make_arg_type_abbr_spine}
\title{Make argument type abbreviation spine}
\usage{
make_arg_type_abbr_spine(argument_nm_1L_chr, object_type_lup)
}
\arguments{
\item{argument_nm_1L_chr}{Argument name (a character vector of length one)}

\item{object_type_lup}{Object type (a lookup table)}
}
\value{
Argument type abbreviation spine (a character vector of length one)
}
\description{
make_arg_type_abbr_spine() is a Make function that creates a new R object Specifically, this function implements an algorithm to make argument type abbreviation spine. The function returns Argument type abbreviation spine (a character vector of length one).
}
\keyword{internal}
