% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_type_lup_ls}
\alias{make_arg_type_lup_ls}
\title{Make argument type lookup table list}
\usage{
make_arg_type_lup_ls(
  object_type_lup = NULL,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  server_1L_chr = deprecated()
)
}
\arguments{
\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Lookup table list (a list of lookup tables)
}
\description{
make_arg_type_lup_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make argument type lookup table list. The function returns Lookup table list (a list of lookup tables).
}
\keyword{internal}
