% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arg_type_lup_ls}
\alias{make_arg_type_lup_ls}
\title{Make argument type}
\usage{
make_arg_type_lup_ls(object_type_lup = NULL)
}
\arguments{
\item{object_type_lup}{Object type (a lookup table), Default: NULL}
}
\value{
Lookup table list (a list of lookup tables)
}
\description{
make_arg_type_lup_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make argument type lookup table list. The function returns a lookup table list (a list of lookup tables).
}
\keyword{internal}
