% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dmt_for_all_fns}
\alias{make_dmt_for_all_fns}
\title{Make documentation for all functions}
\usage{
make_dmt_for_all_fns(
  paths_ls = make_fn_nms(),
  undocumented_fns_dir_chr = make_undmtd_fns_dir_chr(),
  custom_dmt_ls = list(details_ls = NULL, inc_for_main_user_lgl_ls =
    list(force_true_chr = NA_character_, force_false_chr = NA_character_), args_ls_ls =
    NULL),
  fn_type_lup_tb,
  abbreviations_lup = NULL,
  inc_all_mthds_1L_lgl = T
)
}
\arguments{
\item{paths_ls}{Paths (a list), Default: make_fn_nms()}

\item{undocumented_fns_dir_chr}{Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr()}

\item{custom_dmt_ls}{Custom documentation (a list), Default: list(details_ls = NULL, inc_for_main_user_lgl_ls = list(force_true_chr = NA_character_,
force_false_chr = NA_character_), args_ls_ls = NULL)}

\item{fn_type_lup_tb}{Function type lookup table (a tibble)}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{inc_all_mthds_1L_lgl}{Include all methods (a logical vector of length one), Default: T}
}
\value{
All functions documentation (a tibble)
}
\description{
make_dmt_for_all_fns() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make documentation for all functions. The function returns All functions documentation (a tibble).
}
