% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_desc}
\alias{make_fn_desc}
\title{Make function description}
\usage{
make_fn_desc(
  fns_chr,
  title_chr,
  output_chr,
  fns_env_ls,
  fn_types_lup = NULL,
  abbreviations_lup = NULL,
  test_for_write_R_warning_fn = NULL,
  is_generic_lgl = F
)
}
\arguments{
\item{fns_chr}{Functions (a character vector)}

\item{title_chr}{Title (a character vector)}

\item{output_chr}{Output (a character vector)}

\item{fns_env_ls}{Functions (a list of environments)}

\item{fn_types_lup}{Function types (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{test_for_write_R_warning_fn}{Test for write warning (a function), Default: NULL}

\item{is_generic_lgl}{Is generic (a logical vector), Default: F}
}
\value{
Function description (a character vector)
}
\description{
make_fn_desc() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function description. The function returns Function description (a character vector).
}
\keyword{internal}
