% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_dmt_tbl}
\alias{make_fn_dmt_tbl}
\title{Make function documentation table}
\usage{
make_fn_dmt_tbl(
  fns_path_chr,
  fns_dir_chr = make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T),
  custom_dmt_ls = list(title_ls = NULL, desc_ls = NULL, details_ls = NULL,
    inc_for_main_user_lgl_ls = NULL, output_ls = NULL, example_ls = NULL, args_ls_ls =
    NULL),
  append_1L_lgl = T,
  fn_type_lup_tb = NULL,
  abbreviations_lup = NULL,
  dv_ds_nm_1L_chr = "https://doi.org/10.7910/DVN/2Y9VF9",
  dv_url_pfx_1L_chr = NULL,
  key_1L_chr = NULL,
  object_type_lup = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  test_for_write_R_warning_fn = NULL
)
}
\arguments{
\item{fns_path_chr}{Functions path (a character vector)}

\item{fns_dir_chr}{Functions directory (a character vector), Default: make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)}

\item{custom_dmt_ls}{Custom documentation (a list), Default: list(title_ls = NULL, desc_ls = NULL, details_ls = NULL, inc_for_main_user_lgl_ls = NULL,
output_ls = NULL, example_ls = NULL, args_ls_ls = NULL)}

\item{append_1L_lgl}{Append (a logical vector of length one), Default: T}

\item{fn_type_lup_tb}{Function type lookup table (a tibble), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'https://doi.org/10.7910/DVN/2Y9VF9'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{test_for_write_R_warning_fn}{Test for write warning (a function), Default: NULL}
}
\value{
Function documentation table (a tibble)
}
\description{
make_fn_dmt_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function documentation table. The function returns Function documentation table (a tibble).
}
\keyword{internal}
