% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_dmt_tbl}
\alias{make_fn_dmt_tbl}
\title{Make function documentation table}
\usage{
make_fn_dmt_tbl(
  fns_path_chr,
  fns_dir_chr = make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T),
  custom_dmt_ls = make_custom_dmt_ls(),
  append_1L_lgl = T,
  fns_env_ls = NULL,
  fn_types_lup = NULL,
  abbreviations_lup = NULL,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  object_type_lup = NULL,
  server_1L_chr = deprecated(),
  test_for_write_R_warning_fn = NULL
)
}
\arguments{
\item{fns_path_chr}{Functions path (a character vector)}

\item{fns_dir_chr}{Functions directory (a character vector), Default: make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)}

\item{custom_dmt_ls}{Custom documentation (a list), Default: make_custom_dmt_ls()}

\item{append_1L_lgl}{Append (a logical vector of length one), Default: T}

\item{fns_env_ls}{Functions (a list of environments), Default: NULL}

\item{fn_types_lup}{Function types (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}

\item{test_for_write_R_warning_fn}{Test for write warning (a function), Default: NULL}
}
\value{
Function documentation table (a tibble)
}
\description{
make_fn_dmt_tbl() is a Make function that creates a new R object Specifically, this function implements an algorithm to make function documentation table. The function returns Function documentation table (a tibble).
}
\keyword{internal}
