% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_title}
\alias{make_fn_title}
\title{Make function title}
\usage{
make_fn_title(
  fns_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  fn_types_lup = NULL,
  is_generic_lgl = F,
  key_1L_chr = deprecated(),
  server_1L_chr = deprecated()
)
}
\arguments{
\item{fns_chr}{Functions (a character vector)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{fn_types_lup}{Function types (a lookup table), Default: NULL}

\item{is_generic_lgl}{Is generic (a logical vector), Default: F}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\value{
Title (a character vector)
}
\description{
make_fn_title() is a Make function that creates a new R object Specifically, this function implements an algorithm to make function title. The function returns Title (a character vector).
}
\keyword{internal}
