% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_type_lup_tb}
\alias{make_fn_type_lup_tb}
\title{Make function type lookup table}
\usage{
make_fn_type_lup_tb()
}
\value{
Function type lookup table (a tibble)
}
\description{
make_fn_type_lup_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a function type a lookup table.The function returns a function type lookup table (a tibble).
}
\keyword{internal}
