% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_obj_lup_spine}
\alias{make_obj_lup_spine}
\title{Make object lookup table spine}
\usage{
make_obj_lup_spine(
  seed_obj_lup_tb = get_rds_from_dv("seed_obj_lup_tb"),
  new_entries_tb = NULL
)
}
\arguments{
\item{seed_obj_lup_tb}{Seed object lookup table (a tibble), Default: get_rds_from_dv("seed_obj_lup_tb")}

\item{new_entries_tb}{New entries (a tibble), Default: NULL}
}
\value{
NA ()
}
\description{
make_obj_lup_spine() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make object lookup table spine. The function is called for its side effects and does not return a value.
}
