% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_obj_lup_spine}
\alias{make_obj_lup_spine}
\title{Make object lookup table spine}
\usage{
make_obj_lup_spine(
  seed_obj_type_lup = get_rds_from_pkg_dmt(fl_nm_1L_chr = "seed_obj_type_lup",
    piggyback_to_1L_chr = "ready4-dev/ready4"),
  new_entries_tb = NULL
)
}
\arguments{
\item{seed_obj_type_lup}{Seed object type (a lookup table), Default: get_rds_from_pkg_dmt(fl_nm_1L_chr = "seed_obj_type_lup", piggyback_to_1L_chr = "ready4-dev/ready4")}

\item{new_entries_tb}{New entries (a tibble), Default: NULL}
}
\value{
obj_lup_spine (An object)
}
\description{
make_obj_lup_spine() is a Make function that creates a new R object Specifically, this function implements an algorithm to make object lookup table spine. The function is called for its side effects and does not return a value.
}
\keyword{internal}
