% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_executor.R
\name{make_pt_ready4fun_executor}
\alias{make_pt_ready4fun_executor}
\title{Make prototype ready4fun package function list ready4 S3 class for declaring package description file data.}
\usage{
make_pt_ready4fun_executor(args_ls = list(), fn = identity)
}
\arguments{
\item{args_ls}{Arguments (a list), Default: list()}

\item{fn}{Function (a function), Default: identity}
}
\value{
A prototype for ready4 S3 class for declaring package description file data.
}
\description{
Create a new prototype for the ready4 S3 class for declaring package description file data.
}
\details{
ready4 S3 class for list object specifying function arguments and function.
}
