% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_object_type_lup.R
\docType{data}
\name{object_type_lup}
\alias{object_type_lup}
\title{Object abbreviations lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 133 rows and 3 columns.
}
\source{
\url{https://ready4-dev.github.io/ready4/}
}
\usage{
object_type_lup
}
\description{
A lookup table to identify R object types from an abbreviation that can be used as object name suffices.
}
\details{
A tibble

\describe{
\item{short_name_chr}{Short name (a character vector)}
\item{long_name_chr}{Long name (a character vector)}
\item{plural_lgl}{Plural (a logical vector)}
}
}
\keyword{datasets}
