% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure.ready4fun_manifest}
\alias{procure.ready4fun_manifest}
\alias{procure,ready4fun_manifest-method}
\title{Procure data by executing a search and retrieval algorithm using data contained in an instance of a class}
\usage{
procure.ready4fun_manifest(x, type_1L_chr)

\S4method{procure}{ready4fun_manifest}(x, type_1L_chr)
}
\arguments{
\item{x}{An instance of ready4 S3 class for encapsulating the metadata required for package set-up.}

\item{type_1L_chr}{Type (a character vector of length one)}
}
\value{
Value (an output object of multiple potential types)
}
\description{
procure.ready4fun_manifest() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function returns Value (an output object of multiple potential types).
}
