% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_read.R
\name{read_fns}
\alias{read_fns}
\title{Read functions}
\usage{
read_fns(fns_dir_1L_chr = "data-raw/fns/", fns_env = NULL)
}
\arguments{
\item{fns_dir_1L_chr}{Functions directory (a character vector of length one), Default: 'data-raw/fns/'}

\item{fns_env}{Functions (an environment), Default: NULL}
}
\value{
Functions (a list of environments)
}
\description{
read_fns() is a Read function that reads an R script into memory. Specifically, this function implements an algorithm to read functions. Function argument fns_dir_1L_chr specifies the path to object. The function returns Functions (a list of environments).
}
\keyword{internal}
