% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4fun.R
\docType{package}
\name{ready4fun-package}
\alias{ready4fun}
\alias{ready4fun-package}
\title{ready4fun: Readyforwhatsnext Function Authoring And Documentation
Tools}
\description{
ready4fun is a collection of functions for
authoring code libraries of functions and datasets for use in mental
health simulations developed within the readyforwhatsnext
ecosystem.
}
\details{
To learn more about ready4fun, start with the vignettes:
\code{browseVignettes(package = "ready4fun")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://readyforwhatsnext.github.io/ready4fun/}
  \item \url{https://github.com/readyforwhatsnext/ready4fun}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
