% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_dataset.R
\name{ready4fun_dataset}
\alias{ready4fun_dataset}
\title{ready4 S3 class for declaring package description file data..1}
\usage{
ready4fun_dataset(x = make_pt_ready4fun_dataset())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for declaring package description file data..1, Default: make_pt_ready4fun_dataset()}
}
\value{
A validated instance of the ready4 S3 class for declaring package description file data..1
}
\description{
Create a new valid instance of the ready4 S3 class for declaring package description file data..1
}
\details{
ready4 S3 class for declaring package description file data.
}
