% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_executor.R
\name{ready4fun_executor}
\alias{ready4fun_executor}
\title{ready4 S3 class for list object specifying function arguments and function.}
\usage{
ready4fun_executor(x = make_pt_ready4fun_executor())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for list object specifying function arguments and function., Default: make_pt_ready4fun_executor()}
}
\value{
A validated instance of the ready4 S3 class for list object specifying function arguments and function.
}
\description{
Create a new valid instance of the ready4 S3 class for list object specifying function arguments and function.
}
\details{
ready4 S3 class for list object specifying function arguments and function.
}
