% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_obj_type_from_nm}
\alias{remove_obj_type_from_nm}
\title{Remove object type from name}
\usage{
remove_obj_type_from_nm(
  nms_chr,
  object_type_lup = NULL,
  abbreviations_lup = NULL,
  is_generic_lgl = F
)
}
\arguments{
\item{nms_chr}{Names (a character vector)}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{is_generic_lgl}{Is generic (a logical vector), Default: F}
}
\value{
Names (a character vector)
}
\description{
remove_obj_type_from_nm() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove object type from name. Function argument nms_chr specifies the object to be updated. Argument object_type_lup provides the object to be updated. The function returns Names (a character vector).
}
