% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew.ready4fun_abbreviations}
\alias{renew.ready4fun_abbreviations}
\alias{renew,ready4fun_abbreviations-method}
\alias{renew.ready4fun_functions}
\alias{renew,ready4fun_functions-method}
\alias{renew.ready4fun_manifest}
\alias{renew,ready4fun_manifest-method}
\alias{renew.ready4fun_objects}
\alias{renew,ready4fun_objects-method}
\title{Renew values in a dataset}
\usage{
renew.ready4fun_abbreviations(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  plural_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

\S4method{renew}{ready4fun_abbreviations}(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  plural_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

renew.ready4fun_functions(
  x,
  fn_type_nm_chr = NA_character_,
  fn_type_desc_chr = NA_character_,
  first_arg_desc_chr = NA_character_,
  second_arg_desc_chr = NA_character_,
  is_generic_lgl = NA,
  is_method_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

\S4method{renew}{ready4fun_functions}(
  x,
  fn_type_nm_chr = NA_character_,
  fn_type_desc_chr = NA_character_,
  first_arg_desc_chr = NA_character_,
  second_arg_desc_chr = NA_character_,
  is_generic_lgl = NA,
  is_method_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

renew.ready4fun_manifest(
  x,
  type_1L_chr,
  append_1L_lgl = F,
  are_words_chr = character(0),
  custom_plural_ls = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  long_name_chr = character(0),
  no_plural_chr = character(0),
  not_obj_type_chr = character(0),
  pfx_rgx = character(0),
  tf_to_singular_chr = character(0)
)

\S4method{renew}{ready4fun_manifest}(
  x,
  type_1L_chr,
  append_1L_lgl = F,
  are_words_chr = character(0),
  custom_plural_ls = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  long_name_chr = character(0),
  no_plural_chr = character(0),
  not_obj_type_chr = character(0),
  pfx_rgx = character(0),
  tf_to_singular_chr = character(0)
)

renew.ready4fun_objects(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  atomic_element_lgl = NA,
  r3_can_extend_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

\S4method{renew}{ready4fun_objects}(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  atomic_element_lgl = NA,
  r3_can_extend_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)
}
\arguments{
\item{x}{An instance of Object types lookup table.}

\item{short_name_chr}{Short name (a character vector), Default: 'NA'}

\item{long_name_chr}{Long name (a character vector), Default: 'NA'}

\item{plural_lgl}{Plural (a logical vector), Default: NA}

\item{filter_cdn_1L_chr}{Filter condition (a character vector of length one), Default: 'NA'}

\item{new_cases_r3}{New cases (a ready4 S3), Default: NULL}

\item{slice_indcs_int}{Slice indices (an integer vector), Default: NA}

\item{fn_type_nm_chr}{Function type name (a character vector), Default: 'NA'}

\item{fn_type_desc_chr}{Function type description (a character vector), Default: 'NA'}

\item{first_arg_desc_chr}{First argument description (a character vector), Default: 'NA'}

\item{second_arg_desc_chr}{Second argument description (a character vector), Default: 'NA'}

\item{is_generic_lgl}{Is generic (a logical vector), Default: NA}

\item{is_method_lgl}{Is method (a logical vector), Default: NA}

\item{type_1L_chr}{Type (a character vector of length one)}

\item{append_1L_lgl}{Append (a logical vector of length one), Default: F}

\item{are_words_chr}{Are words (a character vector), Default: character(0)}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{no_plural_chr}{No plural (a character vector), Default: character(0)}

\item{not_obj_type_chr}{Not object type (a character vector), Default: character(0)}

\item{pfx_rgx}{Prefix (a regular expression vector), Default: character(0)}

\item{tf_to_singular_chr}{Transform to singular (a character vector), Default: character(0)}

\item{atomic_element_lgl}{Atomic element (a logical vector), Default: NA}

\item{r3_can_extend_lgl}{Ready4 S3 can extend (a logical vector), Default: NA}
}
\value{
x (An object)

x (An object)

x (An object)

x (An object)
}
\description{
renew.ready4fun_abbreviations() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 S3 class for tibble object lookup table of abbreviations. The function is called for its side effects and does not return a value.

renew.ready4fun_functions() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the Function types lookup table. The function is called for its side effects and does not return a value.

renew.ready4fun_manifest() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.

renew.ready4fun_objects() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the Object types lookup table. The function is called for its side effects and does not return a value.
}
