% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew.ready4fun_manifest}
\alias{renew.ready4fun_manifest}
\alias{renew,ready4fun_manifest-method}
\title{Renew an instance of a class by updating it with new data}
\usage{
renew.ready4fun_manifest(
  x,
  type_1L_chr,
  are_words_chr = character(0),
  custom_plural_ls = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  long_name_chr = character(0),
  no_plural_chr = character(0),
  not_obj_type_chr = character(0),
  pfx_rgx = character(0),
  tf_to_singular_chr = character(0)
)

\S4method{renew}{ready4fun_manifest}(
  x,
  type_1L_chr,
  are_words_chr = character(0),
  custom_plural_ls = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  long_name_chr = character(0),
  no_plural_chr = character(0),
  not_obj_type_chr = character(0),
  pfx_rgx = character(0),
  tf_to_singular_chr = character(0)
)
}
\arguments{
\item{x}{An instance of ready4 S3 class for encapsulating the metadata required for package set-up.}

\item{type_1L_chr}{Type (a character vector of length one)}

\item{are_words_chr}{Are words (a character vector), Default: character(0)}

\item{custom_plural_ls}{Custom plural (a list), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{long_name_chr}{Long name (a character vector), Default: character(0)}

\item{no_plural_chr}{No plural (a character vector), Default: character(0)}

\item{not_obj_type_chr}{Not object type (a character vector), Default: character(0)}

\item{pfx_rgx}{Prefix (a regular expression vector), Default: character(0)}

\item{tf_to_singular_chr}{Transform to singular (a character vector), Default: character(0)}
}
\value{
x (An object)
}
\description{
renew.ready4fun_manifest() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
