% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_report.R
\name{report}
\alias{report}
\alias{report.ready4fun_manifest}
\alias{report,ready4fun_manifest-method}
\title{Report}
\usage{
report(x, ...)

<<<<<<< HEAD
report.ready4fun_manifest(x, dv_url_pfx_1L_chr = NULL, key_1L_chr = NULL)

\S4method{report}{ready4fun_manifest}(x, dv_url_pfx_1L_chr = NULL, key_1L_chr = NULL)
=======
report.ready4fun_manifest(
  x,
  dv_url_pfx_1L_chr = character(0),
  key_1L_chr = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)

\S4method{report}{ready4fun_manifest}(
  x,
  dv_url_pfx_1L_chr = character(0),
  key_1L_chr = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
>>>>>>> dev
}
\arguments{
\item{x}{An instance of ready4 S3 class for encapsulating the metadata required for package set-up.}

\item{...}{Additional arguments (an additional arguments)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: character(0)}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\description{
report() is a Report function that authors a report. Specifically, this function implements an algorithm to report. The function is called for its side effects and does not return a value.

report.ready4fun_manifest() is a Report method that authors a report. This method is implemented for the ready4 S3 class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.
}
