% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_fn_dmt_with_slots}
\alias{update_fn_dmt_with_slots}
\title{Update function documentation with slots}
\usage{
update_fn_dmt_with_slots(fn_name_1L_chr, fn_dmt_1L_chr)
}
\arguments{
\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{fn_dmt_1L_chr}{Function documentation (a character vector of length one)}
}
\value{
Function documentation (a character vector of length one)
}
\description{
update_fn_dmt_with_slots() is an Update function that edits an object, while preserving core object attributes Specifically, this function implements an algorithm to update function documentation with slots. The function returns Function documentation (a character vector of length one).
}
\keyword{internal}
