% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{validate}
\alias{validate}
\title{Validate}
\usage{
validate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
validate() is a Validate function that validates that an object conforms to required criteria. Specifically, this function implements an algorithm to ratify. The function is called for its side effects and does not return a value.
}
