% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_validate.R
\name{validate_pkg_setup}
\alias{validate_pkg_setup}
\title{Validate package setup}
\usage{
validate_pkg_setup(pkg_setup_ls)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}
}
\value{
Package setup (a list)
}
\description{
validate_pkg_setup() is a Validate function that validates that an object conforms to required criteria. Specifically, this function implements an algorithm to validate package setup. The function returns Package setup (a list).
}
\keyword{internal}
