% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4fun_description.R
\name{validate_ready4fun_description}
\alias{validate_ready4fun_description}
\title{validate ready4fun package description ready4 submodule class for declaring package description file data.}
\usage{
validate_ready4fun_description(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 submodule class for declaring package description file data.}
}
\value{
A prototpe for ready4 submodule class for declaring package description file data.
}
\description{
Validate an instance of the ready4 submodule class for declaring package description file data.
}
\details{
Package description metadata.
}
\keyword{internal}
