% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_all_fn_dmt}
\alias{write_all_fn_dmt}
\title{Write all function documentation}
\usage{
write_all_fn_dmt(pkg_setup_ls, fns_dmt_tb, fns_env_ls, document_unexp_lgl = F)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{fns_dmt_tb}{Functions documentation (a tibble)}

\item{fns_env_ls}{Functions (a list of environments)}

\item{document_unexp_lgl}{Document unexported (a logical vector), Default: F}
}
\description{
write_all_fn_dmt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write all function documentation. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
