% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_and_doc_fn_fls}
\alias{write_and_doc_fn_fls}
\title{write and document function files}
\usage{
write_and_doc_fn_fls(
  pkg_setup_ls,
  make_pdfs_1L_lgl = T,
  update_pkgdown_1L_lgl = T,
  list_generics_1L_lgl = F,
  dev_pkgs_chr = deprecated(),
  fns_dmt_tb = deprecated(),
  path_to_dmt_dir_1L_chr = deprecated(),
  path_to_dvpr_dmt_dir_1L_chr = deprecated(),
  path_to_pkg_rt_1L_chr = deprecated(),
  path_to_user_dmt_dir_1L_chr = deprecated(),
  r_dir_1L_chr = deprecated()
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{make_pdfs_1L_lgl}{Make pdfs (a logical vector of length one), Default: T}

\item{update_pkgdown_1L_lgl}{Update pkgdown (a logical vector of length one), Default: T}

\item{list_generics_1L_lgl}{List generics (a logical vector of length one), Default: F}

\item{dev_pkgs_chr}{Development packages (a character vector), Default: deprecated()}

\item{fns_dmt_tb}{Functions documentation (a tibble), Default: deprecated()}

\item{path_to_dmt_dir_1L_chr}{Path to documentation directory (a character vector of length one), Default: deprecated()}

\item{path_to_dvpr_dmt_dir_1L_chr}{Path to developer documentation directory (a character vector of length one), Default: deprecated()}

\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: deprecated()}

\item{path_to_user_dmt_dir_1L_chr}{Path to user documentation directory (a character vector of length one), Default: deprecated()}

\item{r_dir_1L_chr}{R directory (a character vector of length one), Default: deprecated()}
}
\description{
write_and_doc_fn_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write and document function files. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
