% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_clss}
\alias{write_clss}
\title{Write classes}
\usage{
write_clss(
  pkg_setup_ls,
  key_1L_chr = NULL,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL,
  cls_fn_ls = deprecated(),
  dv_url_pfx_1L_chr = deprecated(),
  dss_records_ls = deprecated()
)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{self_serve_fn_ls}{Self serve (a list of functions), Default: NULL}

\item{cls_fn_ls}{Class (a list of functions), Default: deprecated()}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{dss_records_ls}{Datasets records (a list), Default: deprecated()}
}
\description{
write_clss() is a Write Classes function that writes new classes. Specifically, this function implements an algorithm to write classes. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
