% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_dmtd_fn_type_lup}
\alias{write_dmtd_fn_type_lup}
\title{Write documented function type}
\usage{
write_dmtd_fn_type_lup(
  fn_type_lup_tb = make_fn_type_lup(),
  overwrite_1L_lgl = T,
  pkg_nm_1L_chr = get_dev_pkg_nm(),
  url_1L_chr = url_1L_chr,
  abbreviations_lup = NULL
)
}
\arguments{
\item{fn_type_lup_tb}{Function type lookup table (a tibble), Default: make_fn_type_lup()}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: T}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: get_dev_pkg_nm()}

\item{url_1L_chr}{Url (a character vector of length one), Default: url_1L_chr}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\description{
write_dmtd_fn_type_lup() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to Write documented function type lookup table.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
