% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ds_dmt}
\alias{write_ds_dmt}
\title{Write dataset documentation}
\usage{
write_ds_dmt(
  db_df,
  db_1L_chr,
  title_1L_chr,
  desc_1L_chr,
  format_1L_chr = "A tibble",
  url_1L_chr = NA_character_,
  vars_ls = NULL,
  R_dir_1L_chr = "R",
  simple_lup_1L_lgl = F,
  abbreviations_lup = NULL,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  object_type_lup = NULL,
  server_1L_chr = deprecated()
)
}
\arguments{
\item{db_df}{Database (a data.frame)}

\item{db_1L_chr}{Database (a character vector of length one)}

\item{title_1L_chr}{Title (a character vector of length one)}

\item{desc_1L_chr}{Description (a character vector of length one)}

\item{format_1L_chr}{Format (a character vector of length one), Default: 'A tibble'}

\item{url_1L_chr}{Url (a character vector of length one), Default: 'NA'}

\item{vars_ls}{Variables (a list), Default: NULL}

\item{R_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}

\item{simple_lup_1L_lgl}{Simple lookup table (a logical vector of length one), Default: F}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: deprecated()}

\item{key_1L_chr}{Key (a character vector of length one), Default: deprecated()}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: deprecated()}
}
\description{
write_ds_dmt() is a Write function that writes a file to a specified local directory Specifically, this function implements an algorithm to write dataset documentation. The function is called for its side effects and does not return a value.
}
\keyword{internal}
