% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fn_dmt}
\alias{write_fn_dmt}
\title{Write function documentation}
\usage{
write_fn_dmt(
  fn_name_1L_chr,
  fn_type_1L_chr,
  fn = NULL,
  fn_desc_1L_chr = NA_character_,
  fn_out_type_1L_chr = NA_character_,
  fn_title_1L_chr = NA_character_,
  example_1L_lgl = F,
  export_1L_lgl = T,
  class_name_1L_chr = "",
  details_1L_chr = "DETAILS",
  args_ls = NULL,
  import_chr = NA_character_,
  doc_in_class_1L_lgl = F,
  abbreviations_lup = NULL,
  object_type_lup = NULL
)
}
\arguments{
\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{fn_type_1L_chr}{Function type (a character vector of length one)}

\item{fn}{Function (a function), Default: NULL}

\item{fn_desc_1L_chr}{Function description (a character vector of length one), Default: 'NA'}

\item{fn_out_type_1L_chr}{Function out type (a character vector of length one), Default: 'NA'}

\item{fn_title_1L_chr}{Function title (a character vector of length one), Default: 'NA'}

\item{example_1L_lgl}{Example (a logical vector of length one), Default: F}

\item{export_1L_lgl}{Export (a logical vector of length one), Default: T}

\item{class_name_1L_chr}{Class name (a character vector of length one), Default: ''}

\item{details_1L_chr}{Details (a character vector of length one), Default: 'DETAILS'}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{import_chr}{Import (a character vector), Default: 'NA'}

\item{doc_in_class_1L_lgl}{Document in class (a logical vector of length one), Default: F}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}
}
\description{
write_fn_dmt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a function documentation.NA
}
\keyword{internal}
