% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fn_fl}
\alias{write_fn_fl}
\title{Write function file}
\usage{
write_fn_fl(
  fns_env_ls,
  pkg_setup_ls,
  document_unexp_lgl = T,
  consent_1L_chr = NULL,
  fns_dmt_tb = deprecated(),
  r_dir_1L_chr = deprecated()
)
}
\arguments{
\item{fns_env_ls}{Functions (a list of environments)}

\item{pkg_setup_ls}{Package setup (a list)}

\item{document_unexp_lgl}{Document unexported (a logical vector), Default: T}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: NULL}

\item{fns_dmt_tb}{Functions documentation (a tibble), Default: deprecated()}

\item{r_dir_1L_chr}{R directory (a character vector of length one), Default: deprecated()}
}
\description{
write_fn_fl() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write function file. The function is called for its side effects and does not return a value.
}
\keyword{internal}
