% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_manuals_to_dv}
\alias{write_manuals_to_dv}
\title{Write manuals to dataverse}
\usage{
write_manuals_to_dv(
  package_1L_chr = get_dev_pkg_nm(getwd()),
  path_to_dmt_dir_1L_chr,
  pkg_dmt_dv_ds_1L_chr,
  publish_dv_1L_lgl = F
)
}
\arguments{
\item{package_1L_chr}{Package (a character vector of length one), Default: get_dev_pkg_nm(getwd())}

\item{path_to_dmt_dir_1L_chr}{Path to documentation directory (a character vector of length one)}

\item{pkg_dmt_dv_ds_1L_chr}{Package documentation dataverse dataset (a character vector of length one)}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: F}
}
\description{
write_manuals_to_dv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write manuals to dataverse. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
