% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_package}
\alias{write_package}
\title{Write package}
\usage{
write_package(
  pkg_desc_ls,
  pkg_ds_ls_ls,
  pkg_setup_ls,
  dv_url_pfx_1L_chr = "https://dataverse.harvard.edu/api/access/datafile/",
  path_to_dmt_dir_1L_chr = normalizePath("../../../../../Documentation/Code"),
  publish_dv_1L_lgl = F
)
}
\arguments{
\item{pkg_desc_ls}{Package description (a list)}

\item{pkg_ds_ls_ls}{Package dataset (a list of lists)}

\item{pkg_setup_ls}{Package setup (a list)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: 'https://dataverse.harvard.edu/api/access/datafile/'}

\item{path_to_dmt_dir_1L_chr}{Path to documentation directory (a character vector of length one), Default: normalizePath("../../../../../Documentation/Code")}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: F}
}
\description{
write_package() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write package. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
