% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_pkg_setup_fls}
\alias{write_pkg_setup_fls}
\title{Write package setup files}
\usage{
write_pkg_setup_fls(
  path_to_pkg_rt_1L_chr = ".",
  dev_pkg_nm_1L_chr = get_dev_pkg_nm(),
  make_tmpl_vignette_1L_lgl = F,
  incr_ver_1L_lgl = T,
  delete_contents_of_R_dir = F
)
}
\arguments{
\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: '.'}

\item{dev_pkg_nm_1L_chr}{Development package name (a character vector of length one), Default: get_dev_pkg_nm()}

\item{make_tmpl_vignette_1L_lgl}{Make tmpl vignette (a logical vector of length one), Default: F}

\item{incr_ver_1L_lgl}{Incr ver (a logical vector of length one), Default: T}

\item{delete_contents_of_R_dir}{PARAM_DESCRIPTION, Default: F}
}
\description{
write_pkg_setup_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to Write package setup files.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
