% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_pkg_setup_fls}
\alias{write_pkg_setup_fls}
\title{write package setup files}
\usage{
write_pkg_setup_fls(
  pkg_desc_ls,
  copyright_holders_chr,
  gh_repo_1L_chr,
  addl_badges_ls = list(),
  badges_lup = NULL,
  check_type_1L_chr = "none",
  delete_r_dir_cnts_1L_lgl = F,
  lifecycle_stage_1L_chr = "experimental",
  incr_ver_1L_lgl = T,
  on_cran_1L_lgl = F,
  path_to_pkg_logo_1L_chr = NA_character_,
  add_gh_site_1L_lgl = T,
  dev_pkg_nm_1L_chr = get_dev_pkg_nm(getwd()),
  path_to_pkg_rt_1L_chr = getwd()
)
}
\arguments{
\item{pkg_desc_ls}{Package description (a list)}

\item{copyright_holders_chr}{Copyright holders (a character vector)}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one)}

\item{addl_badges_ls}{Additional badges (a list), Default: list()}

\item{badges_lup}{Badges (a lookup table), Default: NULL}

\item{check_type_1L_chr}{Check type (a character vector of length one), Default: 'none'}

\item{delete_r_dir_cnts_1L_lgl}{Delete r directory contents (a logical vector of length one), Default: F}

\item{lifecycle_stage_1L_chr}{Lifecycle stage (a character vector of length one), Default: 'experimental'}

\item{incr_ver_1L_lgl}{Increment version (a logical vector of length one), Default: T}

\item{on_cran_1L_lgl}{On cran (a logical vector of length one), Default: F}

\item{path_to_pkg_logo_1L_chr}{Path to package logo (a character vector of length one), Default: 'NA'}

\item{add_gh_site_1L_lgl}{Add github site (a logical vector of length one), Default: T}

\item{dev_pkg_nm_1L_chr}{Development package name (a character vector of length one), Default: get_dev_pkg_nm(getwd())}

\item{path_to_pkg_rt_1L_chr}{Path to package root (a character vector of length one), Default: getwd()}
}
\description{
write_pkg_setup_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write package setup files. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
